\name{AirDens}
\alias{AirDens}
\title{Air Density}
\usage{
  AirDens(pres, T, rv)
}
\arguments{
  \item{pres}{air pressure (pascals)}

  \item{T}{air temperature (Kelvin)}

  \item{rv}{humidity}
}
\description{
  estimate air density from pressure, temperature, and
  humidity
}
\author{
  Mike Dietze
}

