% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_met.R
\name{align.met}
\alias{align.met}
\title{align.met}
\usage{
align.met(train.path, source.path, yrs.train = NULL, n.ens = NULL,
  pair.mems = TRUE, seed = Sys.Date(), verbose = FALSE)
}
\arguments{
\item{train.path}{- path to the dataset to be used to downscale the data}

\item{source.path}{- data to be bias-corrected aligned with training data (from align.met)}

\item{yrs.train}{- (optional) specify a specific years to be loaded for the training data; 
prevents needing to load the entire dataset.  If NULL, all available years 
will be loaded. If not null, should be a vector of numbers (so you can skip
problematic years)}

\item{n.ens}{- number of ensemble members to generate and save}

\item{pair.mems}{- logical stating whether ensemble members should be paired}

\item{seed}{- specify seed so that random draws can be reproduced}
}
\description{
This script aligns meteorology datasets in at temporal resolution for debiasing & 
             temporal downscaling.  
             Note: The output here is stored in memory!  
             Note: can probably at borrow from or adapt align_data.R in Benchmarking module, but 
             it's too much of a black box at the moment.
}
\details{
Align meteorology datasets for debiasing
}
\seealso{
Other debias - Debias & Align Meteorology Datasets into continuous time series: \code{\link{debias.met.regression}}
}
\author{
Christy Rollinson
}
