% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_ERA5.R
\name{Average_ERA5_2_GeoTIFF}
\alias{Average_ERA5_2_GeoTIFF}
\title{Average_ERA5_2_GeoTIFF}
\usage{
Average_ERA5_2_GeoTIFF(start.date, end.date, in.path, outdir)
}
\arguments{
\item{start.date}{character: start point of when to average the data (e.g., 2012-01-01).}

\item{end.date}{character: end point of when to average the data (e.g., 2021-12-31).}

\item{in.path}{character: the directory where your ERA5 data stored (they should named as ERA5_YEAR.nc).}

\item{outdir}{character: the output directory where the averaged GeoTIFF file will be generated.}
}
\value{
character: path to the exported GeoTIFF file.
}
\description{
This function helps to average the ERA5 data based on the start and end dates, and convert it to the GeoTIFF file.
}
\author{
Dongchen Zhang
}
