% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.downscaling.R
\name{cfmet.downscale.daily}
\alias{cfmet.downscale.daily}
\title{Simple, Fast Daily to Hourly Climate Downscaling}
\usage{
cfmet.downscale.daily(dailymet, output.dt = 1, lat)
}
\arguments{
\item{dailymet}{data frame with climate variables}

\item{output.dt}{output timestep}

\item{lat}{latitude (for calculating solar radiation)}
}
\value{
weather file with subdaily timesteps
}
\description{
Based on weach family of functions but 5x faster than weachNEW,
and requiring metric units (temperature in Kelvins on input and celsius on
  output, windspeed in kph, precip in mm, relative humidity as fraction).
Derived from the weachDT function in the BioCro package.
}
\author{
David LeBauer
}
