% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.downscaling.R
\name{cfmet.downscale.time}
\alias{cfmet.downscale.time}
\title{Temporal downscaling of daily or subdaily CF met data}
\usage{
cfmet.downscale.time(cfmet, output.dt = 1, lat = lat, ...)
}
\arguments{
\item{cfmet}{data frame with CF variables generated by \code{\link{load.cfmet}}}

\item{output.dt}{time step (hours) for output}

\item{lat}{latitude (for calculating solar radiation)}

\item{...}{ignored}
}
\value{
downscaled result
}
\description{
Temporal downscaling of daily or subdaily CF met data
}
\author{
David LeBauer
}
