% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cfmet.temporal.downscale}
\alias{cfmet.temporal.downscale}
\title{Downscale CF met data}
\usage{
cfmet.temporal.downscale(cfmet, lat, output.dt = 1)
}
\arguments{
\item{cfmet}{data.table with CF variables generated by \code{\link{load.cfmet}}}

\item{output.dt}{time step for output}
}
\value{
downscaled result
}
\description{
Temporal downscaling of daily or subdaily met data
}
\author{
David LeBauer
}

