% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_has_required_variable}
\alias{check_has_required_variable}
\title{Check that a required variable is present in the given met input
NetCDF}
\usage{
check_has_required_variable(nc_vars, required_var)
}
\arguments{
\item{nc_vars}{Character vector of NetCDF variable names to check}

\item{required_var}{Character vector if required variables.}
}
\value{
Error message if absent, or `TRUE` if present.
}
\description{
Check that a required variable is present in the given met input
NetCDF
}
\author{
Alexey Shiklomanov
}
