% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_message}
\alias{check_message}
\title{Execute a message statement based on}
\usage{
check_message(check)
}
\arguments{
\item{check}{Character string saying}
}
\value{
Executes a logger statement dependant on
}
\description{
Execute a message statement based on
}
\author{
Tony Gardella
}
