% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_met_input_file}
\alias{check_met_input_file}
\title{Check a meteorology data file for compliance with the PEcAn standard}
\usage{
check_met_input_file(metfile, throw_error = TRUE,
  variable_table = pecan_standard_met_table,
  required_vars = pecan_standard_met_table \%>\% dplyr::filter(is_required)
  \%>\% dplyr::pull(cf_standard_name))
}
\arguments{
\item{metfile}{Path of met file to check, as a scalar character.}

\item{throw_error}{Logical. If `TRUE`, throw an informative error
if any tests fail. If `FALSE`,}

\item{variable_table}{`data.frame` linking standard names to their
units. Defaults to [pecan_standard_met_table].}
}
\value{
Tidy `data.frame` summarizing the results of the tests.
}
\description{
Check a meteorology data file for compliance with the PEcAn standard
}
\author{
Alexey Shiklomanov
}
