% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_met_variable}
\alias{check_met_variable}
\title{Check a met variable for PEcAn standard compliance}
\usage{
check_met_variable(nc, variable, variable_table = pecan_standard_met_table,
  warn_unknown = TRUE)
}
\arguments{
\item{nc}{NetCDF object (as returned by [ncdf4::nc_open]).}

\item{variable}{Variable to check, as a character scalar.}

\item{variable_table}{`data.frame` linking standard names to their
units. Defaults to [pecan_standard_met_table].}

\item{warn_unknown}{Logical. If `TRUE`, warn about unknown
variables.}
}
\value{
`TRUE` (invisibly) if successful, otherwise throw an informative error.
}
\description{
Checks that the variable has a correct CF standard name, and the
correct units.
}
\author{
Alexey Shiklomanov
}
