% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_unit}
\alias{check_unit}
\title{Check that the unit of a variable in a NetCDF file is equivalent to
the expected unit.}
\usage{
check_unit(variable, nc, variable_table, warn_unknown = TRUE)
}
\arguments{
\item{variable}{Name of target variable, as a length 1 character}

\item{nc}{NetCDF object containing target variable}

\item{variable_table}{`data.frame` linking standard names to their
units. Must contain columns "cf_standard_name" and "units".
Default is [pecan_standard_met_table].}

\item{warn_unknown}{Logical. If `TRUE` (default), throw a warning
for variables not in `variable_table`. Otherwise, ignore unknown
variables.}
}
\value{
`TRUE` if unit is correct, or `try-error` object if there is a mismatch.
}
\description{
Check that the unit of a variable in a NetCDF file is equivalent to
the expected unit.
}
\author{
Alexey Shiklomanov
}
