% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{col2ncvar}
\alias{col2ncvar}
\title{Create `ncvar` object from variable name}
\usage{
col2ncvar(variable, dims)
}
\arguments{
\item{variable}{CF variable name}

\item{dims}{List of NetCDF dimension objects (passed to 
`ncdf4::ncvar_def(..., dim)`)}
}
\value{
`ncvar` object (from `ncvar_def`)
}
\description{
Create `ncvar` object from variable name
}
