% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyvals.R
\name{copyvals}
\alias{copyvals}
\title{copyvals}
\usage{
copyvals(nc1, var1, nc2, var2, dim2, units2 = NA, conv = NULL,
  missval = -6999, verbose = FALSE)
}
\arguments{
\item{verbose=FALSE}{should ouput of function be extra verbose}
}
\description{
Get meteorology variables from Ameriflux L2 netCDF files and convert to netCDF CF format
}
\author{
Josh Mantooth, Mike Dietze, Elizabeth Cowdery, Ankur Desai
}

