% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.process.R
\name{db.site.lat.lon}
\alias{db.site.lat.lon}
\title{Look up lat/lon from siteid}
\usage{
db.site.lat.lon(site.id, con)
}
\arguments{
\item{site.id}{BeTY ID of site to look up}

\item{con}{database connection}
}
\description{
Look up lat/lon from siteid
}
\author{
Betsy Cowdery
}
