% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debias.met.R
\name{debias.met}
\alias{debias.met}
\title{debias.met}
\usage{
debias.met(outfolder, source_met, train_met, site_id, de_method = "linear",
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{source_met}{- the met data that you want to be altered}

\item{train_met}{- the met data that we use to train the other dataset}

\item{de_method}{- select which debias method you would like to use, options are 'normal', 'linear regression'}
}
\description{
debias.met takes source_met, train_met and output new_met that debiases source_met against train_met
}
\author{
James Simkins
}

