% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debias_met_regression.R
\name{debias.met.regression}
\alias{debias.met.regression}
\title{debias.met.regression}
\usage{
debias.met.regression(train.data, source.data, n.ens, vars.debias = NULL,
  CRUNCEP = FALSE, pair.anoms = TRUE, pair.ens = FALSE,
  uncert.prop = "mean", resids = FALSE, seed = Sys.Date(), outfolder,
  yrs.save = NULL, ens.name, ens.mems = NULL, force.sanity = TRUE,
  sanity.tries = 25, lat.in, lon.in, save.diagnostics = TRUE,
  path.diagnostics = NULL, parallel = FALSE, n.cores = NULL,
  overwrite = TRUE, verbose = FALSE)
}
\arguments{
\item{train.data}{- training data coming out of align.met}

\item{source.data}{- data to be bias-corrected aligned with training data (from align.met)}

\item{n.ens}{- number of ensemble members to generate and save for EACH source ensemble member}

\item{vars.debias}{- which met variables should be debiased? if NULL, all variables in train.data}

\item{CRUNCEP}{- flag for if the dataset being downscaled is CRUNCEP; if TRUE, special cases triggered for 
met variables that have been naively gapfilled for certain time periods}

\item{pair.anoms}{- logical stating whether anomalies from the same year should be matched or not}

\item{pair.ens}{- logical stating whether ensembles from train and source data need to be paired together 
(for uncertainty propogation)}

\item{uncert.prop}{- method for error propogation if only 1 ensemble member; options=c(random, mean); *Not Implemented yet}

\item{resids}{- logical stating whether to pass on residual data or not *Not implemented yet}

\item{seed}{- specify seed so that random draws can be reproduced}

\item{outfolder}{- directory where the data should go}

\item{yrs.save}{- what years from the source data should be saved; if NULL all years of the source data will be saved}

\item{ens.name}{- what is the name that should be attached to the debiased ensemble}

\item{ens.mems}{- what labels/numbers to attach to the ensemble members so we can gradually build bigger ensembles 
without having to do do giant runs at once; if NULL will be numbered 1:n.ens}

\item{force.sanity}{- (logical) do we force the data to meet sanity checks?}

\item{sanity.tries}{- how many time should we try to predict a reasonable value before giving up?  We don't want to end up in an infinite loop}

\item{lat.in}{- latitude of site}

\item{lon.in}{- longitude of site}

\item{save.diagnostics}{- logical; save diagnostic plots of output?}

\item{path.diagnostics}{- path to where the diagnostic graphs should be saved}

\item{parallel}{- (experimental) logical stating whether to run temporal_downscale_functions.R in parallel *Not Implemented yet}

\item{n.cores}{- (experimental) how many cores to use in parallelization *Not implemented yet}

\item{overwrite}{- overwrite existing files? Currently ignored}

\item{verbose}{logical: should \code{\link[ncdf4:ncdf4-package]{ncdf4}}
functions print debugging information as they run?}
}
\description{
This script debiases one dataset (e.g. GCM, re-analysis product) given another higher 
             resolution product or empirical observations. It assumes input are in annual CF standard 
             files that are generate from the pecan extract or download funcitons.
}
\details{
Debias Meteorology using Multiple Linear Regression
Statistically debias met datasets and generate ensembles based on the observed uncertainty
}
\seealso{
Other debias - Debias & Align Meteorology Datasets into continuous time series: \code{\link{align.met}}
}
\author{
Christy Rollinson
}
\concept{debias - Debias & Align Meteorology Datasets into continuous time series}
