% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.caladapt.loca.R
\name{download.caladapt.loca}
\alias{download.caladapt.loca}
\title{Retrieve Cal-Adapt climate raster for a polygon or a point}
\usage{
download.caladapt.loca(
  outfolder,
  start_date,
  end_date,
  sf_obj,
  var,
  gcm,
  scenario,
  period
)
}
\arguments{
\item{outfolder}{Character. Output directory path to store downloaded data. (Required, no default)}

\item{start_date}{Numeric. Start date (year) for data retrieval. Accepts YYYY-MM-DD format, but caladaptr functions only use the year.}

\item{end_date}{Numeric. End date (year) for data retrieval. Accepts YYYY-MM-DD format, but caladaptr functions only use the year.}

\item{sf_obj}{An `sf`, `sfc`, or `SpatVector` object representing the points or polygon boundary.}

\item{var}{Character. Climate variable to retrieve. One of "tasmax", "tasmin", "pr",
"swe", "baseflow", "et", "rainfall",
"runoff", "snowfall", "soilMoist1", "Tair".}

\item{gcm}{Character. GCM name. One of "HadGEM2-ES", "CNRM-CM5", "CanESM2", "MIROC5", "ACCESS1-0",
"CCSM4", "CESM1-BGC", "CMCC-CMS", "GFDL-CM3", "HadGEM2-CC", "ens32avg",
"ens32max", "ens32min".}

\item{scenario}{Character. Emission scenario. One of "historical", "rcp45", "rcp85".}

\item{period}{Character. Time aggregation. One of "day", "month", "year", "30yavg".}
}
\value{
A tibble row with appended climate raster in the `raster` column.
For points, the row will also include `lat` and `lon` columns.
}
\description{
Fetches climate raster data from Cal-Adapt for a specified polygon or point,
GCM, emission scenario, and time period. Supports LOCA (CMIP5-based) data,
and ensures that the geometry is within the bounding box of valid coordinates
for California.
}
\examples{
\dontrun{
 library(caladaptr)
 yolo_polygon <- caladaptr::ca_aoipreset_geom("counties") |>
   dplyr::filter(state_name == "California") |>
   dplyr::filter(name == "Yolo") |>
   dplyr::select(state_name, county_name = name, geom) |>
   sf::st_transform(4326)
 polygon_rasters <- download.caladapt.loca(
   outfolder = tempdir(),
   start_date = 2006, end_date = 2100,
   sf_obj = yolo_polygon,
   var = "pr",
   gcm = "CNRM-CM5",
   scenario = "rcp85",
   period = "year"
 )
 z <- stars::read_stars(polygon_rasters$raster)
 plot(z)
 # There is also a caladaptr function ca_stars_read
 # that returns a list of stars objects.
 coords <- data.frame(lat = 36, lon = -120) |>
   sf::st_as_sf(coords = c("lon", "lat"), crs = 4326)
 point_rasters <- download.caladapt.loca(
   outfolder = tempdir(),
   start_date = 2006, end_date = 2100,
   sf_obj = coords,
   var = "pr",
   gcm = "CNRM-CM5",
   scenario = "rcp85",
   period = "year"
 )
}
}
\references{
- [Cal-Adapt](https://caladapt.org/)
- [Cal Adapt Data Catalog Documentation](https://berkeley-gif.github.io/caladapt-docs/data-catalog.html)
- [Cal-Adapt Loca Data Catalog](https://albers.cnr.berkeley.edu/data/scripps/loca/met/)
}
