\name{download.CRUNCEP}
\alias{download.CRUNCEP}
\title{download.CRUNCEP}
\usage{
download.CRUNCEP(outfolder, start_date, end_date, site_id, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
  \item{outfolder}{}

  \item{start_date}{}

  \item{end_date}{}

  \item{lat}{}

  \item{lon}{}
}
\description{
Download and conver to CF CRUNCEP single grid point from
MSTIMIP server using OPENDAP interface
}
\author{
James Simkins, Mike Dietze
}

