% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.Drought2018.R
\name{download.Drought2018}
\alias{download.Drought2018}
\title{Download ICOS Drought 2018 data}
\usage{
download.Drought2018(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE
)
}
\arguments{
\item{sitename}{ICOS id of the site. Example - "BE-Bra"}

\item{outfolder}{path to the directory where the output file is stored. If specified directory does not exists, it is created.}

\item{start_date}{start date of the data request in the form YYYY-MM-DD}

\item{end_date}{end date area of the data request in the form YYYY-MM-DD}

\item{overwrite}{should existing files be overwritten. Default False.}
}
\value{
information about the output file
}
\description{
Link to the product: https://www.icos-cp.eu/data-products/YVR0-4898
}
\details{
Variables present in the output CSV file: TA_F, SW_IN_F, LW_IN_F, VPD_F, PA_F, P_F, WS_F, WD, RH, PPFD_IN, CO2_F_MDS, TS_F_MDS_1, TS_F_MDS_2, NEE_VUT_REF, LE_F_MDS, RECO_NT_VUT_REF and GPP_NT_VUT_REF

Compatible with FLUXNET 2015 format
}
\examples{
download.Drought2018("FI-Sii", "/home/carya/pecan", "2016-01-01", "2018-01-01")

}
\author{
Ayush Prasad
}
