% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ECMWF.R
\name{download.ECMWF}
\alias{download.ECMWF}
\title{Downloads ECMWF Open Data 15 day forecast data products 
https://confluence.ecmwf.int/display/DAC/ECMWF+open+data%3A+real-time+forecasts}
\usage{
download.ECMWF(
  outfolder,
  lat.in,
  lon.in,
  fchour = 0,
  product = NULL,
  type = c("cf", "pf"),
  parameters = "all",
  overwrite = FALSE,
  reticulate_python = NULL,
  ...
)
}
\arguments{
\item{outfolder}{location on disk where outputs will be stored}

\item{lat.in, lon.in}{site coordinates, decimal degrees (numeric)}

\item{fchour}{reference time of the forecasts. Values are 00, 12 for 15 day forecast.
`fchour` is not passed via upstream functions and we are using 00 by default when the function is called from within the pecan workflow.}

\item{type}{the type of forecast data. Current values: `cf` (controlled forecast), `pf` (perturbed forecast)}

\item{parameters}{character vector of product types, or `"all"`.}

\item{overwrite}{Logical. If `FALSE` (default), skip any files with
  the same target name (i.e. same variable) that already exist in
  `outfolder`. If `TRUE`, silently overwrite existing files.

Forecast hour is reference time of the forecasts. Values are 00, 12 for 15 day forecast.}

\item{reticulate_python}{Path to Python binary for `reticulate`
(passed to [reticulate::use_python()]). If `NULL` (default), use
the system default.}
}
\value{
information about the output file
}
\description{
https://github.com/ecmwf/ecmwf-opendata
Under the hood, this function uses the Python `ecmwf-opendata` module.
The module and dependencies can be accessed via the `reticulate` package.
}
\details{
`reticulate::conda_install(c("scipy", "xarray", "eccodes", "cfgrib", "ecmwf-opendata"), envname = "r-reticulate", pip = TRUE)`
`reticulate::use_condaenv("r-reticulate")`
}
\author{
Swarnalee Mazumder
}
