% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.FluxnetLaThuile.R
\name{download.FluxnetLaThuile}
\alias{download.FluxnetLaThuile}
\title{download.FluxnetLaThuile}
\usage{
download.FluxnetLaThuile(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  username = "pecan",
  ...
)
}
\arguments{
\item{sitename}{the FLUXNET ID of the site to be downloaded, used as file name prefix.
The 'SITE_ID' field in \href{http://www.fluxdata.org/DataInfo/Dataset\%20Doc\%20Lib/SynthDataSummary.aspx}{list of Fluxnet LaThuile sites}}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{username}{should be the registered Fluxnet username, else defaults to pecan}

\item{...}{further arguments, currently ignored}
}
\description{
Download Flxunet LaThuile CSV files
}
\author{
Ankur Desai
}
