% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.geostreams.R
\name{download.geostreams}
\alias{download.geostreams}
\title{Download Geostreams data from Clowder API}
\usage{
download.geostreams(outfolder, sitename, start_date, end_date,
  url = "https://terraref.ncsa.illinois.edu/clowder/api/geostreams", ...)
}
\arguments{
\item{outfolder}{directory in which to save json result. Will be created if necessary}

\item{sitename}{character. Should match a geostreams sensor_name}

\item{start_date, end_date}{datetime}

\item{url}{base url for Clowder host}

\item{...}{other arguments passed as query parameters}
}
\description{
Download Geostreams data from Clowder API
}
\examples{
\dontrun{
 download.geostreams(outfolder = "~/output/dbfiles/Clowder_EF",
                     sitename = "UIUC Energy Farm - CEN",
                     start_date = "2016-01-01", end_date="2016-12-31",
                     key="verysecret")
}
}
\author{
Harsh Agrawal, Chris Black
}
