% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.GFDL.R
\name{download.GFDL}
\alias{download.GFDL}
\title{download.GFDL}
\usage{
download.GFDL(outfolder, start_date, end_date, site_id, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, model, experiment, scenario, ...)
}
\arguments{
\item{model}{, select which GFDL model to run (options are CM3, ESM2M, ESM2G)}

\item{experiment}{, select which experiment to run (options are rcp26, rcp45, rcp60, rcp85)}

\item{scenario}{, select which scenario to initialize the run (options are r1i1p1, r3i1p1, r5i1p1)}
}
\description{
Download GFDL CMIP5 outputs for a single grid point using OPeNDAP and convert to CF
}
\author{
James Simkins
}

