% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ICOS.R
\name{download.ICOS}
\alias{download.ICOS}
\title{Download ICOS Ecosystem data products}
\usage{
download.ICOS(
  sitename,
  outfolder,
  start_date,
  end_date,
  product,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{sitename}{ICOS id of the site. Example - "BE-Bra"}

\item{outfolder}{path to the directory where the output file is stored. If specified directory does not exists, it is created.}

\item{start_date}{start date of the data request in the form YYYY-MM-DD}

\item{end_date}{end date area of the data request in the form YYYY-MM-DD}

\item{product}{ICOS product to be downloaded. Currently supported options: "Drought2018", "ETC"}

\item{overwrite}{should existing files be overwritten. Default False.}

\item{...}{used when extra arguments are present.}
}
\value{
information about the output file
}
\description{
Currently available products:
Drought-2018 ecosystem eddy covariance flux product https://www.icos-cp.eu/data-products/YVR0-4898
ICOS Final Fully Quality Controlled Observational Data (Level 2) https://www.icos-cp.eu/data-products/ecosystem-release
}
\examples{
\dontrun{
download.ICOS("FI-Sii", "/home/carya/pecan", "2016-01-01", "2018-01-01", product = "Drought2018")
}
}
\author{
Ayush Prasad
}
