% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.MACA.R
\name{download.MACA}
\alias{download.MACA}
\title{download.MACA}
\usage{
download.MACA(outfolder, start_date, end_date, site_id, lat.in, lon.in,
  model = "IPSL-CM5A-LR", scenario = "rcp85",
  ensemble_member = "r1i1p1", overwrite = FALSE, verbose = FALSE,
  ...)
}
\arguments{
\item{start_date}{, of the format "YEAR-01-01 00:00:00"}

\item{end_date}{, of the format "YEAR-12-31 23:59:59"}

\item{model}{, select which MACA model to run (options are BNU-ESM, CNRM-CM5, CSIRO-Mk3-6-0, bcc-csm1-1, bcc-csm1-1-m, CanESM2, GFDL-ESM2M, GFDL-ESM2G, HadGEM2-CC365, HadGEM2-ES365, inmcm4, MIROC5, MIROC-ESM, MIROC-ESM-CHEM, MRI-CGCM3, CCSM4, IPSL-CM5A-LR, IPSL-CM5A-MR, IPSL-CM5B-LR, NorESM1-M)}

\item{scenario}{, select which scenario to run (options are rcp45, rcp85)}

\item{ensemble_member}{, r1i1p1 is the only ensemble member available for this dataset, CCSM4 uses r6i1p1 instead}
}
\description{
Download MACA CMIP5 outputs for a single grid point using OPeNDAP and convert to CF
}
\author{
James Simkins
}
