% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.MERRA.R
\name{download.MERRA}
\alias{download.MERRA}
\title{Download MERRA data}
\usage{
download.MERRA(
  outfolder,
  start_date,
  end_date,
  lat.in,
  lon.in,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{outfolder}{Directory where results should be written}

\item{start_date}{Range of years to retrieve. Format is YYYY-MM-DD,
but only the year portion is used and the resulting files always contain a full year of data.}

\item{end_date}{Range of years to retrieve. Format is YYYY-MM-DD,
but only the year portion is used and the resulting files always contain a full year of data.}

\item{lat.in}{site latitude in decimal degrees}

\item{lon.in}{site longitude in decimal degrees}

\item{overwrite}{logical. Download a fresh version even if a local file with the same name already exists?}

\item{verbose}{logical. Passed on to \code{\link[ncdf4]{ncvar_def}} and \code{\link[ncdf4]{nc_create}}
to control printing of debug info}

\item{...}{Not used -- silently soak up extra arguments from `convert.input`, etc.}
}
\value{
`data.frame` of meteorology data metadata
}
\description{
Download MERRA data
}
\author{
Alexey Shiklomanov
}
