\name{download.MsTMIP_NARR}
\alias{download.MsTMIP_NARR}
\title{download.MsTMIP_NARR}
\usage{
download.MsTMIP_NARR(outfolder, start_date, end_date, site_id, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
  \item{outfolder}{}

  \item{start_date}{}

  \item{end_date}{}

  \item{lat}{}

  \item{lon}{}
}
\description{
Download and conver to CF NARR single grid point from
MSTIMIP server using OPENDAP interface
}
\author{
James Simkins
}

