% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR_csv.R
\name{download.NARR_csv}
\alias{download.NARR_csv}
\title{Download NARR time series as CSV}
\usage{
download.NARR_csv(outfolder, start_date, end_date, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{outfolder}{Target directory for storing output}

\item{start_date}{Start date for met data}

\item{end_date}{End date for met data}

\item{lat.in}{Site latitude coordinate}

\item{lon.in}{Site longitude coordinate}

\item{overwrite}{Overwrite existing files?  Default=FALSE}

\item{verbose}{Turn on verbose output? Default=FALSE}
}
\description{
Download NARR time series as CSV
}
\examples{
download.NARR_csv(tempdir(), "2001-01-01", "2001-01-12", 43.372, -89.907)

}
\author{
Alexey Shiklomanov
}
