% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NEONmet.R
\name{download.NEONmet}
\alias{download.NEONmet}
\title{Download NEON Site Met CSV files}
\usage{
download.NEONmet(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sitename}{the NEON ID of the site to be downloaded, used as file name prefix.
The 4-letter SITE code  in \href{https://www.neonscience.org/science-design/field-sites/list}{list of NEON sites}}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year and month of the date)}

\item{end_date}{the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year and month part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{makes the function output more text}

\item{...}{further arguments, currently ignored}
}
\description{
download.NEONmet
}
\details{
Uses NEON v0 API to download met data from NEON towers and convert to CF NetCDF
}
\examples{
\dontrun{
result <- download.NEONmet('HARV','~/','2017-01-01','2017-01-31',overwrite=TRUE)
}
}
