% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_noaa_gefs_efi.R
\name{download.NOAA_GEFS_EFI}
\alias{download.NOAA_GEFS_EFI}
\title{download.NOAA_GEFS_EFI}
\usage{
download.NOAA_GEFS_EFI(sitename, outfolder, start_date, site.lat, site.lon)
}
\arguments{
\item{sitename}{NEON site name}

\item{outfolder}{filepath to save ensemble member .nc files}

\item{start_date}{start date for met forecast}

\item{site.lat}{site lat}

\item{site.lon}{site lon}
}
\value{
message confirming download complete and location of .nc files
}
\description{
download.NOAA_GEFS_EFI
}
\author{
Alexis Helgeson

Mike Dietze
}
