% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.raw.met.module.R
\name{download.raw.met.module}
\alias{download.raw.met.module}
\alias{.download.raw.met.module}
\title{download.raw.met.module}
\usage{
.download.raw.met.module(
  dir,
  met,
  register,
  machine,
  start_date,
  end_date,
  str_ns,
  con,
  input_met,
  site.id,
  lat.in,
  lon.in,
  host,
  site,
  username,
  overwrite = FALSE,
  dbparms,
  Ens.Flag = FALSE
)
}
\arguments{
\item{dir}{directory to write outputs to}

\item{met}{source included in input_met}

\item{register}{register.xml, provided by met.process}

\item{machine}{machine associated with hostname, provided by met.process}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{str_ns}{substitute for site_id if not provided, provided by met.process}

\item{con}{database connection based on dbparms in met.process}

\item{input_met}{Which data source to process}

\item{site.id}{site id}

\item{lat.in}{site latitude, provided by met.process}

\item{lon.in}{site longitude, provided by met.process}

\item{host}{host info from settings file}

\item{site}{site info from settings file}

\item{username}{database username}

\item{overwrite}{whether to force download.raw.met.module to proceed}

\item{dbparms}{database settings from settings file}

\item{Ens.Flag}{default set to FALSE}
}
\value{
A list of data frames is returned containing information about the data file that can be used to locate it later.  Each
data frame contains information about one file.
}
\description{
download.raw.met.module
}
