% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.US_WCr.R
\name{download.US_WCr}
\alias{download.US_WCr}
\title{download.US-WCr}
\usage{
download.US_WCr(start_date, end_date, timestep = 1)
}
\arguments{
\item{start_date}{Start date/time data should be downloaded for}

\item{end_date}{End date/time data should be downloaded for}

\item{timestep}{How often to take data points from the file.  Must be a multiple of 0.5}
}
\description{
download.US-WCr
}
\section{General Description}{

Obtains data from Ankur Desai's Willow Creek flux tower, and selects certain variables (NEE and LE) to return
Data is returned at the given timestep in the given range.

This data includes information on a number of flux variables.

The timestep parameter is measured in hours, but is then converted to half hours because the data's timestep
is every half hour.
}

\author{
Luke Dramko
}
