% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.US_Wlef.R
\name{download.US_Wlef}
\alias{download.US_Wlef}
\title{download.US_Wlef}
\usage{
download.US_Wlef(start_date, end_date, timestep = 1)
}
\arguments{
\item{start_date}{Start date/time data should be downloaded for}

\item{end_date}{End date/time data should be downloaded for}

\item{timestep}{How often to take data points from the file.  Must be integer}
}
\description{
download.US_Wlef
}
\section{General Description}{

Obtains data from Ankur Desai's WLEF/ Parks Fall flux tower, and selects certain variables (NEE and LE) to return
Data is returned at the given timestep in the given range.

This data includes information on a number of flux variables.
}

\author{
Luke Dramko and K Zarada
}
