% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEFS_helper_functions.R
\name{download_grid}
\alias{download_grid}
\title{Download all requested timepoints of one GEFS ensemble member}
\usage{
download_grid(
  ens_index,
  location,
  directory,
  hours_char,
  cycle,
  base_filename1,
  vars,
  working_directory
)
}
\arguments{
\item{ens_index}{ensemble member as integer (1-31)}

\item{location}{bounding box portion of query,
as a single URL-escaped string}

\item{directory}{server path portion of query,
as a single URL-escaped string
(not to be confused with local output dir -- that's `working_directory`)}

\item{hours_char}{timepoints to retrieve,
as zero-padded strings e.g. `c("000", "384", "840")`}

\item{cycle}{forecast hour to use ("00", "06", "12", or "18")}

\item{base_filename1}{URL onto which to append query string components}

\item{vars}{variable listing component of query,
as a single URL-escaped string}

\item{working_directory}{path on local disk to write output}
}
\description{
Download all requested timepoints of one GEFS ensemble member
}
