% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_noaa_gefs_efi.R
\name{download_NOAA_GEFS_EFI}
\alias{download_NOAA_GEFS_EFI}
\title{download_NOAA_GEFS_EFI}
\usage{
download_NOAA_GEFS_EFI(met.start, site_id, base_dir, site.lat, site.lon)
}
\arguments{
\item{met.start}{start date for met forecast}

\item{site_id}{NEON site name}

\item{base_dir}{filepath to save ensemble member .nc files}

\item{site.lat}{site lat}

\item{site.lon}{site lon}
}
\value{
message confirming download complete and location of .nc files
}
\description{
download_NOAA_GEFS_EFI
}
\author{
Alexis Helgeson
}
