% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.downscaling.R
\name{downscale_one_cfmet_day}
\alias{downscale_one_cfmet_day}
\title{Internal helper to downscale a single row from a daily file}
\usage{
downscale_one_cfmet_day(df, tseq, lat)
}
\arguments{
\item{df}{one row from dailymet}

\item{tseq}{vector of hours at which to estimate}

\item{lat}{latitude}
}
\value{
df with one row for each hour in `tseq`
}
\description{
Internal helper to downscale a single row from a daily file
}
