% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_hour_downscale.R
\name{downscale_ShortWave_to_half_hrly}
\alias{downscale_ShortWave_to_half_hrly}
\title{Downscale shortwave to half hourly}
\usage{
downscale_ShortWave_to_half_hrly(df, lat, lon, hr = 0.5)
}
\arguments{
\item{df}{data frame of variables}

\item{lat}{lat of site}

\item{lon}{long of site}

\item{hr}{hour to downscale to- default is 1}
}
\value{
A dataframe of downscaled state variables

ShortWave.ds
}
\description{
Downscale shortwave to half hourly
}
\author{
Laura Puckett
}
