% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling_helper_functions.R
\name{downscale_ShortWave_to_hrly}
\alias{downscale_ShortWave_to_hrly}
\title{Downscale shortwave to hourly}
\usage{
downscale_ShortWave_to_hrly(df, lat, lon, hr = 1)
}
\arguments{
\item{df, }{data frame of variables}

\item{lat, }{lat of site}

\item{lon, }{long of site}

\item{hr, }{hour to downscale to- default is 1}
}
\value{
A dataframe of downscaled state variables

ShortWave.ds
}
\description{
Downscale shortwave to hourly
}
\author{
Laura Puckett
}
