% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_hour_downscale.R
\name{downscale_solar_geom_halfhour}
\alias{downscale_solar_geom_halfhour}
\title{Calculate potential shortwave radiation}
\usage{
downscale_solar_geom_halfhour(doy, lon, lat)
}
\arguments{
\item{doy, }{day of year in decimal}

\item{lon, }{longitude}

\item{lat, }{latitude}
}
\value{
vector of potential shortwave radiation for each doy
}
\description{
Calculate potential shortwave radiation
}
\author{
Quinn Thomas
}
