% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_hour_downscale.R
\name{downscale_spline_to_half_hrly}
\alias{downscale_spline_to_half_hrly}
\title{Downscale spline to half hourly}
\usage{
downscale_spline_to_half_hrly(df, VarNames, hr = 0.5)
}
\arguments{
\item{df}{dataframe of data to be downscales}

\item{VarNames}{variable names to be downscaled}

\item{hr}{hour to downscale to- default is 0.5}
}
\value{
A dataframe of half hourly downscaled state variables
}
\description{
Downscale spline to half hourly
}
\author{
Laura Puckett
}
