% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_spline_to_hourly.R
\name{downscale_spline_to_hourly}
\alias{downscale_spline_to_hourly}
\title{Downscale spline to hourly}
\usage{
downscale_spline_to_hourly(df, VarNamesStates)
}
\arguments{
\item{df, }{dataframe of data to be downscales}

\item{VarNamesStates, }{names of vars that are state variables}
}
\value{
A dataframe of downscaled state variables
}
\description{
Downscale spline to hourly
}
\author{
Laura Puckett
}
