% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling_helper_functions.R
\name{downscale_spline_to_hrly}
\alias{downscale_spline_to_hrly}
\title{Downscale spline to hourly}
\usage{
downscale_spline_to_hrly(df, VarNames, hr = 1)
}
\arguments{
\item{df, }{dataframe of data to be downscales}

\item{VarNames, }{variable names to be downscaled}

\item{hr, }{hour to downscale to- default is 1}
}
\value{
A dataframe of downscaled state variables
}
\description{
Downscale spline to hourly
}
\author{
Laura Puckett
}
