% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angle.R
\name{eccentricity_obliquity}
\alias{eccentricity_obliquity}
\title{Equation of time: Eccentricity and obliquity}
\usage{
eccentricity_obliquity(doy)
}
\arguments{
\item{doy}{Day of year}
}
\description{
Equation of time: Eccentricity and obliquity
}
\author{
Alexey Shiklomanov
}
