% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angle.R
\name{equation_of_time}
\alias{equation_of_time}
\title{Equation of time: Eccentricity and obliquity}
\usage{
equation_of_time(doy)
}
\arguments{
\item{doy}{Day of year}
}
\value{
`numeric(1)` length of the solar day, in hours.
}
\description{
For description of calculations, see https://en.wikipedia.org/wiki/Equation_of_time#Calculating_the_equation_of_time
}
\author{
Alexey Shiklomanov
}
