% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_download.R
\name{ERA5_cds_annual_download}
\alias{ERA5_cds_annual_download}
\title{ERA5_cds_annual_download}
\usage{
ERA5_cds_annual_download(
  years,
  months,
  days,
  times,
  area,
  variables,
  outdir,
  auto.create.key = T
)
}
\arguments{
\item{years}{Numeric: a series of years to be downloaded (e.g., 2012:2021).}

\item{months}{List: a list contains months to be downloaded (e.g., list("01", "02") to download files in Jan and Feb).}

\item{days}{List: a list contains days to be downloaded (e.g., list("01", "02") to download files in the first and second days).}

\item{times}{List: a list contains times to be downloaded (e.g., list('00:00','03:00') to download files for the times 12:00 and 3:00 am UTC).}

\item{area}{Character: a string contains the bounding box (formatted as "North/West/South/East") to be downloaded (e.g., "85/-179/7/-20").}

\item{variables}{List: a list contains variables to be downloaded (e.g., list("2m_temperature","surface_pressure")).}

\item{outdir}{Character: physical path where the ERA5 data are stored.}

\item{auto.create.key}{Boolean: decide if we want to generate the CDS RC file if it doesn't exist, the default is TRUE.}
}
\value{
A vector containing file paths to the downloaded files.
}
\description{
This function helps to download the yearly ERA5 data based on the prescribed features using the CDS API.
}
\author{
Dongchen Zhang
}
