% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ERA5.R
\name{ERA5_extract}
\alias{ERA5_extract}
\title{ERA5_extract}
\usage{
ERA5_extract(lat = 40, long = -120, years = c(2006), vars = NULL,
  data.folder)
}
\arguments{
\item{lat}{latitude}

\item{long}{longitude}

\item{years}{years to be extracted}

\item{vars}{variables to be extarcted. If NULL all the variables will be returned.}

\item{data.folder}{Path to the directory where ERA5 nc files are located.}
}
\value{
a list of xts objects with all the variables for the requested years
}
\description{
ERA5_extract
}
\details{
For the list of variables check out the documentation at \link{https://confluence.ecmwf.int/display/CKB/ERA5+data+documentation#ERA5datadocumentation-Spatialgrid}
}
\examples{
\dontrun{
point.data <- ERA5_extract(lat=40, long=-120, years=c(1990:1995), vars=NULL, data.folder=".")

 map(~xts::apply.daily(.x, mean))

}
}
