% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extract.nc}
\alias{extract.nc}
\title{extract.nc}
\usage{
extract.nc(in.path, in.prefix, outfolder, start_date, end_date, slat, slon,
  newsite)
}
\description{
Given latitude and longitude coordinates, extract site data from NARR file
}
\author{
Betsy Cowdery
}

