% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{generate_narr_url}
\alias{generate_narr_url}
\title{Generate NARR url from a vector of dates}
\usage{
generate_narr_url(dates, flx)
}
\arguments{
\item{dates}{Vector of dates for which to generate URL}

\item{flx}{(Logical) If `TRUE`, format for `flx` variables. Otherwise, 
format for `sfc` variables. See [narr_flx_vars].}
}
\description{
Figures out file names for the given dates, based on NARR's convoluted and 
inconsistent naming scheme.
}
\author{
Alexey Shiklomanov
}
