% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metutils.R
\name{get.lv}
\alias{get.lv}
\title{Latent heat of vaporization}
\usage{
get.lv(airtemp = 268.6465)
}
\arguments{
\item{airtemp}{air temperature (Kelvin)}
}
\value{
lV   latent heat of vaporization (J kg-1)
}
\description{
calculate latent heat of vaporization for water
}
\author{
Istem Fer
}
