% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temporal.downscaling.R
\name{get.ncvector}
\alias{get.ncvector}
\title{Get time series vector from netCDF file}
\usage{
get.ncvector(var, lati = lati, loni = loni, run.dates = run.dates, met.nc)
}
\arguments{
\item{var}{}

\item{lati}{}

\item{loni}{}

\item{run.dates}{}

\item{met.nc}{netcdf file with CF variable names}
}
\value{
numeric vector
}
\description{
Get time series vector from netCDF file
}
\details{
internal convenience function for
streamlining extraction of data from netCDF files
with CF-compliant variable names
}
\author{
David Shaner LeBauer
}

