% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cf_variables_table.R
\name{get_cf_variables_table}
\alias{get_cf_variables_table}
\title{Retrieve the current CF variables table from `cfconventions.org`
and convert it into a `data.frame`}
\usage{

  get_cf_variables_table(cf_url = "http://cfconventions.org/Data/cf-standard-names/57/src/cf-standard-name-table.xml")
}
\arguments{
\item{cf_url}{URL of CF variables table XML}
}
\value{
CF variables table, as a `tibble`
}
\description{
Retrieve the current CF variables table from `cfconventions.org`
and convert it into a `data.frame`
}
\author{
Alexey Shiklomanov
}
