% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.Geostreams.R
\name{get_clowderauth}
\alias{get_clowderauth}
\title{Authentication lookup helper}
\usage{
get_clowderauth(key, user, pass, url, authfile = "~/.pecan.clowder.xml")
}
\arguments{
\item{key, user, pass}{passed unchanged from \code{\link{download.Geostreams}} call, possibly null}

\item{url}{matched against \code{<hostname>} in authfile, ignored if authfile contains no hostname.}

\item{authfile}{path to a PEcAn-formatted XML settings file; must contain a \code{<clowder>} key}
}
\description{
Authentication lookup helper
}
