% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_thredds.R
\name{get_narr_url}
\alias{get_narr_url}
\title{Retrieve NARR data from a given URL}
\usage{
get_narr_url(url, xy, flx, pb = NULL)
}
\arguments{
\item{url}{Full URL to NARR thredds file}

\item{xy}{Vector length 2 containing NARR coordinates}

\item{flx}{(Logical) If `TRUE`, format for `flx` variables. Otherwise, 
format for `sfc` variables. See [narr_flx_vars].}

\item{pb}{Progress bar R6 object (default = `NULL`)}
}
\description{
Retrieve NARR data from a given URL
}
\author{
Alexey Shiklomanov
}
