% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_hour_downscale.R
\name{half_hour_downscale}
\alias{half_hour_downscale}
\alias{temporal_downscale_half_hour}
\title{half_hour_downscale}
\usage{
temporal_downscale_half_hour(
  input_file,
  output_file,
  overwrite = TRUE,
  hr = 0.5
)
}
\arguments{
\item{input_file}{location of NOAAGEFS_1hr files}

\item{output_file}{location where to store half_hour files}

\item{overwrite}{whether to force hamf_hour_downscale to proceed}

\item{hr}{set half hour}
}
\value{
A list of data frames is returned containing information about the data file that can be used to locate it later.  Each
data frame contains information about one file.
}
\description{
half_hour_downscale
}
