\name{InputForWeach}
\alias{InputForWeach}
\title{InputForWeach}
\usage{
  InputForWeach(lat, lon, year1, year2)
}
\arguments{
  \item{lat}{numeric latitude}

  \item{lon}{numeric longitude}

  \item{year1}{integer}

  \item{year2}{integer}
}
\value{
  climate data to be parsed by the \code{\link{weachNEW}}
  function in BioCro
}
\description{
  Download Met Drivers for BioCro from NCEP
}
\author{
  Deepak Jaiswal, David LeBauer
}

