% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR_csv.R
\name{latlon2lcc}
\alias{latlon2lcc}
\title{Convert latitude and longitude to x-y coordinates (in km) in Lambert 
conformal conic projection (used by NARR)}
\usage{
latlon2lcc(lat.in, lon.in)
}
\arguments{
\item{lat.in}{Latitude coordinate}

\item{lon.in}{Longitude coordinate}
}
\value{
`sp::SpatialPoints` object containing transformed x and y 
coordinates, in km, which should match NARR coordinates
}
\description{
Convert latitude and longitude to x-y coordinates (in km) in Lambert 
conformal conic projection (used by NARR)
}
\author{
Alexey Shiklomanov
}
