% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NARR_csv.R
\name{latlon2narr}
\alias{latlon2narr}
\title{Convert latitude and longitude coordinates to NARR indices}
\usage{
latlon2narr(nc, lat.in, lon.in)
}
\arguments{
\item{nc}{`ncdf4` connection object}

\item{lat.in}{Latitude coordinate}

\item{lon.in}{Longitude coordinate}
}
\value{
Vector length 2 containing NARR `x` and `y` indices, which can be 
used in `ncdf4::ncvar_get` `start` argument.
}
\description{
Convert latitude and longitude coordinates to NARR indices
}
\author{
Alexey Shiklomanov
}
