% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_lm_ensemble_sims.R
\name{lm_ensemble_sims}
\alias{lm_ensemble_sims}
\title{lm_ensemble_sims}
\usage{
lm_ensemble_sims(dat.mod, n.ens, path.model, lags.list = NULL,
  lags.init = NULL, dat.train)
}
\arguments{
\item{dat.mod}{- dataframe to be predicted at the time step of the training data}

\item{n.ens}{- number of hourly ensemble members to generate}

\item{path.model}{- path to where the training model & betas is stored}

\item{lags.init}{- a data frame of initialization parameters to match the data in dat.mod}

\item{dat.train}{- the training data used to fit the model; needed for night/day in 
surface_downwelling_shortwave_flux_in_air}
}
\description{
This function does the heavy lifting in the final
             function of the tdm workflow titled predict_subdaily_met(). It uses a linear
             regression approach by generating the hourly values from the coarse data of
             the file the user selects to downscale based on the hourly models and betas
             generated by gen.subdaily.models().
}
\details{
Linear Regression Ensemble Simulation
Met downscaling function that predicts ensembles of downscaled meteorology
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{model.train}}, \code{\link{nc.merge}},
  \code{\link{predict_subdaily_met}},
  \code{\link{save.betas}}, \code{\link{save.model}},
  \code{\link{subdaily_pred}},
  \code{\link{temporal.downscale.functions}}
}
\author{
Christy Rollinson, James Simkins
}
