% Generated by roxygen2 (4.0.2): do not edit by hand
\name{load.cfmet}
\alias{load.cfmet}
\title{load CF met}
\usage{
load.cfmet(met.nc, lat, lon, start.date, end.date)
}
\arguments{
\item{met.nc}{object of class ncdf4 representing an open CF compliant, PEcAn standard netcdf file with met data}

\item{lat}{numeric value of latutude}

\item{lon}{numeric value of longitude}

\item{start.date}{format is "YYYY-MM-DD"}

\item{end.date}{format is "YYYY-MM-DD"}
}
\value{
data.table of met data
}
\description{
Load met data from PEcAn formatted met driver
}
\details{
subsets a PEcAn formatted met driver file and converts to a data.table / data.frame object
}
\author{
David LeBauer
}

