% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2CF.ALMA.R
\name{met2CF.ALMA}
\alias{met2CF.ALMA}
\title{met2CF.ALMA}
\usage{
met2CF.ALMA(in.path, in.prefix, outfolder, start_date, end_date,
  overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}
}
\description{
Get meteorology variables from ALMA netCDF files and convert to netCDF CF format
}
\author{
Mike Dietze
}

