% Generated by roxygen2 (4.0.1): do not edit by hand
\name{met2CF.Ameriflux}
\alias{met2CF.Ameriflux}
\title{met2CF.Ameriflux}
\usage{
met2CF.Ameriflux(in.path, in.prefix, outfolder)
}
\arguments{
\item{in.path}{}

\item{in.prefix}{}

\item{outfolder}{}
}
\description{
Get meteorology variables from Ameriflux L2 netCDF files and convert to netCDF CF format
}
\author{
Josh Mantooth, Mike Dietze
}

