% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2CF.csv.R
\name{met2CF.csv}
\alias{met2CF.csv}
\title{Convert met data from CSV to CF}
\usage{
met2CF.csv(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  format,
  lat = NULL,
  lon = NULL,
  nc_verbose = FALSE,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{in.path}{directory in which to find met csv files}

\item{in.prefix}{pattern to match to find met files inside `in.path`}

\item{outfolder}{directory name to write CF outputs}

\item{start_date, end_date}{when to start and stop conversion.
Specify as `Date` objects, but only the year component is used}

\item{format}{data frame or list produced by `PEcAn.DB::query.format.vars`.
See details}

\item{lat, lon}{latitude and longitude of site, in decimal degrees.
If not provided, these are taken from `format`.}

\item{nc_verbose}{logical: run ncvar_add in verbose mode?}

\item{overwrite}{Logical: Redo conversion if output file already exists?}

\item{...}{other arguments, currently ignored}
}
\description{
Convert met data from CSV to CF
}
\details{
The `format` argument takes an output from `PEcAn.DB::query.format.vars`,
   and should have the following components:

REQUIRED:
\itemize{
  \item `format$lat`: latitude of site (unless passed by `lat`)
  \item `format$lon`: longitude of site (unless passed by `lon`)
  \item `format$header`: number of lines of header
  \item `format$vars`: a data.frame with lists of information for each
        variable to read. At least `airT` is required
  \item `format$vars$input_name`: name in CSV file
  \item `format$vars$input_units`: units in CSV file
  \item `format$vars$bety_name`: name in BETY.
      See https://pecan.gitbooks.io/pecan-documentation/content/developers_guide/Adding-an-Input-Converter.html
      for allowable names.
}

OPTIONAL:
\itemize{
  \item `format$na.strings`: list of missing values to convert to NA,
        such as -9999
  \item `format$skip`: lines to skip excluding header
  \item `format$vars$column_number`: column number in CSV file
        (optional, will use header name first)
}

Columns with NA for bety variable name are dropped.

Units for datetime field are the lubridate function that will be used to
   parse the date (e.g. \code{ymd_hms} or \code{mdy_hm}).
}
\examples{
\dontrun{
con <- PEcAn.DB::db.open(
  list(
    user = "bety", password = "bety", host = "localhost",
    dbname = "bety", driver = "PostgreSQL", write = TRUE
  )
)
start_date <- lubridate::ymd_hm("200401010000")
end_date <- lubridate::ymd_hm("200412312330")
file <- PEcAn.data.atmosphere::download.Fluxnet2015("US-WCr", "~/", start_date, end_date)
in.path <- "~/"
in.prefix <- file$dbfile.name
outfolder <- "~/"
format.id <- 5000000001
format <- PEcAn.DB::query.format.vars(format.id = format.id, bety = bety)
format$lon <- -92.0
format$lat <- 45.0
format$time_zone <- "America/Chicago"
results <- PEcAn.data.atmosphere::met2CF.csv(
  in.path, in.prefix, outfolder,
  start_date, end_date, format,
  overwrite = TRUE
)
}

}
\author{
Mike Dietze, David LeBauer, Ankur Desai
}
