% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2cf.ERA5.R
\name{met2cf.ERA5}
\alias{met2cf.ERA5}
\title{met2cf.ERA5}
\usage{
met2cf.ERA5(lat, long, years, sitename, data.folder, outfolder,
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{lat}{latitude}

\item{long}{longitude}

\item{years}{years to be extracted}

\item{sitename}{The name of the site used for making the identifier.}

\item{data.folder}{Path to the directory where ERA5 nc files are located.}

\item{outfolder}{Path to directory where nc files need to be saved.}

\item{overwrite}{Logical flag definign if the files need to be overwriten in case they were downloaded before.}

\item{verbose}{Logical flag defining if ouput of function be extra verbose.}
}
\description{
met2cf.ERA5
}
\examples{
\dontrun{
files.era5 <- met2cf.ERA5(years=2018,
                  data.folder=".",
                  outfolder = "./metfiles")

}
}
