% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2cf.ERA5.reanalysis.R
\name{met2CF.ERA5.reanalysis}
\alias{met2CF.ERA5.reanalysis}
\title{met2CF.ERA5.reanalysis - function for ERA5 reanalysis data processing}
\usage{
met2CF.ERA5.reanalysis(
  slat,
  slon,
  in.path,
  start_date,
  end_date,
  sitename,
  outfolder,
  in.prefix,
  vars = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{slat}{latitude}

\item{slon}{longitude}

\item{in.path}{path to the directory containing the ERA5 reanalysis NetCDF files}

\item{start_date}{start date}

\item{end_date}{end date}

\item{sitename}{the name of the site used for making the identifier}

\item{outfolder}{path to directory where CF-compliant nc files need to be saved}

\item{in.prefix}{initial portion of the filename that does not vary by date}

\item{vars}{variables to be extracted. If NULL all available variables will be returned}

\item{overwrite}{logical if files need to be overwritten}

\item{verbose}{logical flag defining if output of function be extra verbose}
}
\value{
dataframe with file information for processed NetCDF files
}
\description{
met2CF.ERA5.reanalysis - function for ERA5 reanalysis data processing
}
\author{
Akash
}
