% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2CF.FACE.R
\name{met2CF.FACE}
\alias{met2CF.FACE}
\title{convert FACE files to CF files}
\usage{
met2CF.FACE(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  input.id,
  site,
  format,
  ...
)
}
\arguments{
\item{in.path}{directory in which to find FACE files}

\item{in.prefix}{pattern to match to find FACE files inside `in.path`}

\item{outfolder}{directory name to write CF outputs}
}
\description{
convert FACE files to CF files
}
\author{
Elizabeth Cowdery
}
