% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2CF.Geostreams.R
\name{met2CF.Geostreams}
\alias{met2CF.Geostreams}
\title{Convert geostreams JSON to CF met file}
\usage{
met2CF.Geostreams(in.path, in.prefix, outfolder, start_date, end_date,
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{in.path}{directory containing Geostreams JSON file(s) to be converted}

\item{in.prefix}{initial portion of input filenames (everything before the dates)}

\item{outfolder}{directory where nc output files should be written. Will be created if necessary}

\item{start_date, end_date}{beginning and end of run, YYYY-MM-DD.}

\item{overwrite}{logical: Regenerate existing files of the same name?}

\item{verbose}{logical, passed on to \code{\link[ncdf4]{nc_create}} 
to control how chatty it should be during netCDF creation}

\item{...}{other arguments, currently ignored}
}
\description{
Convert geostreams JSON to CF met file
}
\author{
Harsh Agrawal, Chris Black
}
