% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.module.R
\name{met2model.module}
\alias{met2model.module}
\title{met2model.module}
\usage{
met2model.module(ready.id, model, con, host, dir, met, str_ns, site, start_date,
  end_date, browndog, new.site, overwrite = FALSE, exact.dates)
}
\arguments{
\item{ready.id}{}

\item{model}{model_type name}

\item{con}{}

\item{host}{Host info from settings file}

\item{dir}{directory to write outputs to}

\item{met}{Which data source to process.}

\item{site}{Site info from settings file}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{overwrite}{Whether to force met.process to proceed.}

\item{dbparms}{database settings from settings file}
}
\author{
Elizabeth Cowdery, Michael Dietze, Ankur Desai, James Simkins, Ryan Kelly
}
