% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metgapfill.NOAA_GEFS.R
\name{metgapfill.NOAA_GEFS}
\alias{metgapfill.NOAA_GEFS}
\title{Gapfill NOAA_GEFS weather data}
\usage{
metgapfill.NOAA_GEFS(
  in.prefix,
  in.path,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{in.prefix}{the met file name}

\item{in.path}{The location of the file}

\item{outfolder}{The place to write the output file to}

\item{start_date}{The start date of the contents of the file}

\item{end_date}{The end date of the contents of the file}

\item{overwrite}{Whether or not to overwrite the output file if it exists or not}

\item{verbose}{Passed to nc writing functions for additional output}

\item{...}{further arguments, currently ignored}
}
\description{
This function uses simple methods to gapfill NOAA GEFS met data.
Temperature and Precipitation are gapfilled with splines;
 other data sources are gapfilled using linear models fitted to
 other fitted data.
}
\author{
Luke Dramko
}
