% Generated by roxygen2 (4.0.1): do not edit by hand
\name{metgapfill}
\alias{metgapfill}
\title{Take an Ameriflux NetCDF file
Fill missing met values using MDS approach using MPI-BGC REddyProc library
Currently
Future version: Choose which variables to gap fill
Future version will first downscale and fill with NARR, then REddyProc}
\usage{
metgapfill(in.path, in.prefix, outfolder)
}
\arguments{
\item{in.path}{}

\item{in.prefix}{}

\item{outfolder}{}
}
\description{
Take an Ameriflux NetCDF file
Fill missing met values using MDS approach using MPI-BGC REddyProc library
Currently
Future version: Choose which variables to gap fill
Future version will first downscale and fill with NARR, then REddyProc
}
\author{
Ankur Desai
}

