% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_model_train.R
\name{model.train}
\alias{model.train}
\title{model.train}
\usage{
model.train(dat.subset, v, n.beta, resids = resids, threshold = NULL, ...)
}
\arguments{
\item{dat.subset}{data.frame containing lags, next, and downscale period data}

\item{v}{variable name, as character}

\item{n.beta}{number of betas to pull from}

\item{resids}{TRUE or FALSE, whether to use residuals or not}

\item{threshold}{NULL except for surface_downwelling_shortwave_radiation, helps with our
distinction between day and night (no shortwave without sunlight)}

\item{...}{further arguments, currently ignored}
}
\description{
Function to create linear regression models for specific met
             variables. This is used in conjunction with temporal.downscale.functions()
             to generate linear regression statistics and save their output to be called
             later in lm_ensemble_sims().
}
\details{
TDM Model Train
Linear regression calculations for specific met variables
}
\seealso{
Other tdm - Temporally Downscale Meteorology: 
\code{\link{gen.subdaily.models}()},
\code{\link{lm_ensemble_sims}()},
\code{\link{nc.merge}()},
\code{\link{predict_subdaily_met}()},
\code{\link{save.betas}()},
\code{\link{save.model}()},
\code{\link{subdaily_pred}()},
\code{\link{temporal.downscale.functions}()}
}
\author{
Christy Rollinson, James Simkins
}
\concept{tdm - Temporally Downscale Meteorology}
