% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_nc2dat.train.R
\name{nc2dat.train}
\alias{nc2dat.train}
\title{nc2dat.train}
\usage{
nc2dat.train(outfolder, in.path, in.prefix, start_date, end_date,
  upscale = FALSE, overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{outfolder}{- directory where output will be stored}

\item{in.path}{- path of coarse model (e.g. GCM output)}

\item{in.prefix}{- prefix of model string as character (e.g. IPSL.r1i1p1.rcp85)}

\item{start_date}{- yyyy-mm-dd}

\item{end_date}{- yyyy-mm-dd}

\item{upscale}{- Upscale can either be set for FALSE (leave alone) or to the temporal resolution you want to aggregate to}

\item{lat.in}{- latitude as numeric}

\item{lon.in}{- longitude as numeric}
}
\description{
This is the 1st script for the tdm (Temporally Downscale Meteorology) workflow. The nc2dat.train function
             parses multiple netCDF files into one central training data file called 'dat.train_file'. It outputs a 
             netCDF file in CF conventions with a few workflow specific variables (date, doy, and hour) that are used 
             later in the workflow. Thus, this function produces a file unique to the tdm family workflow. This netCDF
             file will be used to generate the subdaily models in the next step of the workflow, tdm_generate_subdaily_models.
             It is also called in tdm_predict_subdaily_met which is the final step of the tdm workflow.
}
\details{
nc2dat.train
Parses multiple netCDF files into one central document for temporal downscaling procedure
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{predict.subdaily.function}},
  \code{\link{predict.subdaily.workflow}},
  \code{\link{temporal.downscale.functions}}
}
\author{
James Simkins, Christy Rollinson
}
