% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEFS_helper_functions.R
\name{noaa_grid_download}
\alias{noaa_grid_download}
\title{Download gridded forecast in the box bounded by the latitude and longitude list}
\usage{
noaa_grid_download(
  lat_list,
  lon_list,
  forecast_time,
  forecast_date,
  model_name_raw,
  output_directory,
  end_hr
)
}
\arguments{
\item{lat_list}{lat for site}

\item{lon_list}{long for site}

\item{forecast_time}{start hour of forecast}

\item{forecast_date}{date for forecast}

\item{model_name_raw}{model name for directory creation}

\item{output_directory}{output directory}

\item{end_hr}{end hr to determine how many hours to download}
}
\value{
NA
}
\description{
Download gridded forecast in the box bounded by the latitude and longitude list
}
