% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noaa_gefs_efi_helper.R
\name{noaa_stage2}
\alias{noaa_stage2}
\title{noaa_stage2}
\usage{
noaa_stage2(
  cycle = 0,
  version = "v12",
  endpoint = "data.ecoforecast.org",
  verbose = TRUE,
  start_date = ""
)
}
\arguments{
\item{cycle}{Hour at which forecast was made, as character string
(`"00"`, `"06"`, `"12"` or `"18"`). Only `"00"` (default) has 30 days horizon.}

\item{version}{GEFS forecast version. Prior versions correspond to forecasts
issued before 2020-09-25 which have different ensemble number and horizon,
among other changes, and are not made available here. Leave as default.}

\item{endpoint}{the EFI host address (leave as default)}

\item{verbose}{logical, displays or hides messages}

\item{start_date}{forecast start date yyyy-mm-dd format}
}
\description{
noaa_stage2
}
\author{
Alexis Helgeson (taken from neon4cast package)
}
