% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan_standard_met_table.R
\docType{data}
\name{pecan_standard_met_table}
\alias{pecan_standard_met_table}
\title{Conversion table for PEcAn standard meteorology}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 10 columns.
}
\usage{
pecan_standard_met_table
}
\description{
Conversion table for PEcAn standard meteorology
}
\keyword{datasets}
