% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{post_process}
\alias{post_process}
\title{Post process raw NARR downloaded data frame}
\usage{
post_process(dat)
}
\arguments{
\item{dat}{Nested `tibble` from mapped call to [get_narr_url]}
}
\description{
Post process raw NARR downloaded data frame
}
