% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_predict_subdaily_workflow.R
\name{predict.subdaily.workflow}
\alias{predict.subdaily.workflow}
\title{predict.subdaily.workflow}
\usage{
\method{predict}{subdaily.workflow}(outfolder, in.path, in.prefix,
  lm.models.base, dat.train_file, start_date, end_date, tdf_file, td_file,
  site.name, lat.in, lon.in, cores.max = 12, ens.hr = 3, n.day = 1,
  resids = FALSE, parallel = FALSE, n.cores = NULL, overwrite = FALSE,
  verbose = FALSE)
}
\arguments{
\item{outfolder}{- directory where output file will be stored}

\item{in.path}{- path to model dataset you wish to temporally downscale}

\item{in.prefix}{- prefix of model dataset, i.e. if file is GFDL.CM3.rcp45.r1i1p1.2006 the prefix is 'GFDL.CM3.rcp45.r1i1p1'}

\item{lm.models.base}{- path to linear regression model folder from 3_gen_subdaily}

\item{dat.train_file}{- location of train_data file}

\item{start_date}{- yyyy-mm-dd}

\item{end_date}{- yyyy-mm-dd}

\item{site.name}{- name of location i.e. Ameriflux Site as character string}

\item{lat.in}{- latitude as numeric}

\item{lon.in}{- longitude as numeric}

\item{cores.max}{- 12}

\item{ens.hr}{- integer selecting number of hourly ensemble members}

\item{n.day}{- 1 # Number of daily ensemble members to process, must be 1 because not passing in daily ensemble}

\item{resids}{- logical stating whether to pass on residual data or not}

\item{parallel}{- logical stating whether to run temporal_downscale_functions.R in parallel}

\item{n.cores}{- deals with parallelization}
}
\description{
This is the final script of the tdm family workflow. This script predicts subdaily meteorology
             from daily means using a linear regression modeling approach. It takes a dataset with 
             daily resolution and temporally downscales it to hourly resolution using the statistics
             generated by tdm_generate_subdaily_model.R. It references the predict.subdaily.function
             located in tdm_predict_subdaily_functon.R which uses a linear regression based approach to downscale.
             We generate multiple ensembles of possible hourly values dictated from the models and betas 
             generated in tdm_generate_subdaily_model.R. Each ensemble member is saved as a netCDF file
             in CF conventions and these files are ready to be used in the general PEcAn workflow.
}
\details{
Predict Subdaily Meteorology Workflow
Predict Subdaily Meteorology based off of statistics created in tdm_generate_subdaily_models.R
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{nc2dat.train}},
  \code{\link{predict.subdaily.function}},
  \code{\link{temporal.downscale.functions}}
}
\author{
Christy Rollinson, James Simkins
}
