% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{prepare_narr_year}
\alias{prepare_narr_year}
\title{Write NetCDF file for a single year of data}
\usage{
prepare_narr_year(dat, file, lat_nc, lon_nc, verbose = FALSE)
}
\arguments{
\item{dat}{NARR tabular data for a single year ([get_NARR_thredds])}

\item{file}{Full path to target file}

\item{lat_nc}{`ncdim` object for latitude}

\item{lon_nc}{`ncdim` object for longitude}

\item{verbose}{}
}
\value{
List of NetCDF variables in data. Creates NetCDF file containing 
data as a side effect
}
\description{
Write NetCDF file for a single year of data
}
