% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{process_error_list}
\alias{process_error_list}
\title{Clean up list output of tests mapped across a vector of variables}
\usage{
process_error_list(error_list, var_names)
}
\arguments{
\item{error_list}{List resulting from calling `purrr::map(...,
purrr::safely(some_test))}

\item{var_names}{Character vector to use to name errors in output
list.}
}
\value{
List of error messages, named according to the variable
  that threw the error. If no error messages were thrown, an empty
  list.
}
\description{
Clean up list output of tests mapped across a vector of variables
}
\author{
Alexey Shiklomanov
}
