% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEFS_helper_functions.R
\name{process_gridded_noaa_download}
\alias{process_gridded_noaa_download}
\title{Extract and temporally downscale points from downloaded grid files}
\usage{
process_gridded_noaa_download(
  lat_list,
  lon_list,
  site_id,
  downscale,
  overwrite,
  forecast_date,
  forecast_time,
  model_name,
  model_name_ds,
  model_name_raw,
  output_directory
)
}
\arguments{
\item{lat_list}{lat for site}

\item{lon_list}{lon for site}

\item{site_id}{Unique site_id for file creation}

\item{downscale}{Logical. Default is TRUE. Downscales from 6hr to hourly}

\item{overwrite}{Logical. Default is FALSE. Should exisiting files be overwritten}

\item{forecast_date}{Date for download}

\item{forecast_time}{Time (0,6,12,18) for start of download}

\item{model_name}{Name of model for file name}

\item{model_name_ds}{Name of downscale file name}

\item{model_name_raw}{Name of raw file name}

\item{output_directory}{Output directory}
}
\value{
List
}
\description{
Extract and temporally downscale points from downloaded grid files
}
