% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.register}
\alias{read.register}
\title{read.register}
\usage{
read.register(register.xml, con)
}
\arguments{
\item{register.xml}{path of xml file}

\item{con}{betydb connection}
}
\description{
read.register
}
\author{
Betsy Cowdery
}

