% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{read_narr_var}
\alias{read_narr_var}
\title{Read a specific variable from a NARR NetCDF file}
\usage{
read_narr_var(nc, xy, variable, unit, datatype, pb = NULL)
}
\arguments{
\item{nc}{`ncdf4` connection object}

\item{xy}{Vector length 2 containing NARR coordinates}

\item{variable}{NARR name of variable to retrieve}

\item{unit}{Output unit of variable to retrieve}

\item{datatype}{Type of NARR file -- either `flx` or `sfc`}

\item{pb}{Progress bar R6 object (default = `NULL`)}
}
\description{
Read a specific variable from a NARR NetCDF file
}
\author{
Alexey Shiklomanov
}
