% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustly.R
\name{robustly}
\alias{robustly}
\title{Adverb to try calling a function `n` times before giving up}
\usage{
robustly(.f, n = 10, timeout = 0.2, silent = TRUE)
}
\arguments{
\item{.f}{Function to call.}

\item{n}{Number of attempts to try}

\item{timeout}{Timeout between attempts, in seconds}

\item{silent}{Silence error messages?}
}
\value{
Modified version of input function
}
\description{
Adverb to try calling a function `n` times before giving up
}
\examples{
rlog <- robustly(log, timeout = 0.3)
try(rlog("fail"))
\dontrun{
 nc_openr <- robustly(ncdf4::nc_open, n = 10, timeout = 0.5)
 nc <- nc_openr(url)
 # ...or just call the function directly
 nc <- robustly(ncdf4::nc_open, n = 20)(url)
 # Useful in `purrr` maps
 many_vars <- purrr::map(varnames, robustly(ncdf4::ncvar_get), nc = nc)
}
}
