% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metutils.R
\name{sat_vapor_pressure}
\alias{sat_vapor_pressure}
\alias{t2es}
\alias{get.es}
\title{Saturation vapor pressure (t2es)}
\usage{
sat_vapor_pressure(
  temp,
  temp_units = "degC",
  out_units = "kPa",
  method = c("ClausiusClapeyron")
)

get.es(temp)

t2es(temp, ...)
}
\arguments{
\item{temp}{see \code{sat_vapor_pressure()}}

\item{temp_units}{input temperature units ("degC","K","degF"), default "degC"}

\item{out_units}{output pressure units ("kPa","hPa","Pa","mb"), default "kPa"}

\item{method}{one of "Magnus","ClausiusClapeyron" (default), or "GoffGratch".
See details for references.}
}
\value{
numeric vector in \code{out_units}
}
\description{
Saturation vapor pressure (t2es)
}
\details{
Alias for temperature→e_s convenience. Identical to \code{sat_vapor_pressure()}.
}
\references{
Alduchov, O. A., & Eskridge, R. E. (1996). Improved Magnus Form Approximation of Saturation Vapor Pressure. J. Appl. Meteor.*, 35(4), 601–609. <doi:10.1175/1520-0450(1996)035<0601:IMFAOS>2.0.CO;2>

Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). \strong{Crop evapotranspiration – Guidelines for computing crop water requirements.} FAO Irrigation and Drainage Paper 56.

Goff, J. A., & Gratch, S. (1946). Low-pressure properties of water from −160 to 212F. Trans. ASHVE, 52, 95–122.

WMO (2014) Guide to Instruments and Methods of Observation (WMO-No. 8), ch. 4.
}
\author{
David LeBauer
}
