% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_save_betas.R
\name{save.betas}
\alias{save.betas}
\title{save.betas}
\usage{
save.betas(model.out, betas, outfile)
}
\arguments{
\item{model.out}{list linear regression model output}

\item{betas}{name of the layer of betas to save (e.g. 'betas' or 'betas.resid')}

\item{outfile}{location where output will be stored}
}
\description{
Function to save betas as a .nc file. This is utilized in
             gen.subdaily.models() when linear regression models are created
}
\details{
TDM Save Betas
Saves betas that are calculated during gen.subdaily.models()
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{lm_ensemble_sims}},
  \code{\link{model.train}}, \code{\link{nc.merge}},
  \code{\link{predict_subdaily_met}},
  \code{\link{save.model}}, \code{\link{subdaily_pred}},
  \code{\link{temporal.downscale.functions}}
}
\author{
Christy Rollinson, James Simkins
}
\concept{tdm - Temporally Downscale Meteorology}
