% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.process.R
\name{site_from_tag}
\alias{site_from_tag}
\title{Function to find the site code for a specific tag}
\usage{
site_from_tag(sitename, tag)
}
\arguments{
\item{sitename}{full name of site}

\item{tag}{abbreviated name of site}
}
\description{
Example:
  sitename = 'Rhinelander Aspen FACE Experiment (FACE-RHIN)'
  tag = 'FACE'
  site_from_tag(sitename,tag) = 'RHIN'
Requires that site names be set up specifically with (tag-sitecode) - this may change
}
\author{
Betsy Cowdery
}
