% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angle.R
\name{solar_angle}
\alias{solar_angle}
\title{Calculate solar angle}
\usage{
solar_angle(doy, lat, lon, dt)
}
\arguments{
\item{doy}{Day of year}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{dt}{Timestep}
}
\description{
Calculate solar angle
}
\author{
Alexey Shiklomanov
}
