% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_wind.R
\name{split_wind}
\alias{split_wind}
\title{Split wind_speed into eastward_wind and northward_wind}
\usage{
split_wind(
  in.path,
  in.prefix,
  start_date,
  end_date,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{in.path}{path to original data}

\item{in.prefix}{prefix of original data}

\item{start_date, end_date}{date (or character in a standard date format). Only year component is used.}

\item{overwrite}{logical: replace output file if it already exists?}

\item{verbose}{logical: should \code{\link[ncdf4:ncdf4-package]{ncdf4}} functions print debugging information as they run?}

\item{...}{other arguments, currently ignored}
}
\value{
nothing. TODO: Return data frame summarizing results
}
\description{
Currently modifies the files IN PLACE rather than creating a new copy of the files an a new DB record.
}
\examples{
\dontrun{
in.path <- "~/paleon/PalEONregional_CF_site_1-24047/"
in.prefix <- ""
outfolder <- "~/paleon/metTest/"
start_date <- "0850-01-01"
end_date <- "2010-12-31"
overwrite <- FALSE
verbose <- TRUE

split_wind(in.path, in.prefix, start_date, end_date, merge.file, overwrite, verbose)
}
}
