% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_met.R
\name{step_means}
\alias{step_means}
\title{take mean at fixed intervals along a vector}
\usage{
step_means(x, step)
}
\arguments{
\item{x}{numeric vector}

\item{step}{integer step size}
}
\value{
numeric of length length(x)/step
}
\description{
take mean at fixed intervals along a vector
}
\details{
User should check that length(x) is an even multiple of step
}
