% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_subdaily_pred.R
\name{subdaily_pred}
\alias{subdaily_pred}
\title{subdaily_pred}
\usage{
subdaily_pred(newdata, model.predict, Rbeta, resid.err = FALSE,
  model.resid = NULL, Rbeta.resid = NULL, n.ens)
}
\arguments{
\item{newdata}{dataframe with data to be downscaled}

\item{model.predict}{saved linear regression model}

\item{Rbeta}{matrix with Rbetas from saved linear regression model}

\item{resid.err}{logical, whether to include residual error or not}

\item{model.resid}{data.frame of model residuals}

\item{Rbeta.resid}{data.frame of Rbeta residuals}

\item{n.ens}{number of ensembles to create}
}
\description{
Function to pull objects created in linear regression models
             and are used to predict subdaily meteorology. This function is called in
             lm_ensemble_sims() to downscale a meteorology product.
             Linear regression models are created in gen.subdaily.models()
}
\details{
Subdaily Prediction
Pulls information from linear regression models to predict subdaily meteorology
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{lm_ensemble_sims}},
  \code{\link{model.train}}, \code{\link{nc.merge}},
  \code{\link{predict_subdaily_met}},
  \code{\link{save.betas}}, \code{\link{save.model}},
  \code{\link{temporal.downscale.functions}}
}
\author{
Christy Rollinson, James Simkins
}
\concept{tdm - Temporally Downscale Meteorology}
