% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{summarize_error_list}
\alias{summarize_error_list}
\title{Convert an error list into a tidy `data.frame`}
\usage{
summarize_error_list(error_list, var_names, test_type)
}
\arguments{
\item{error_list}{List of `testthat` results}

\item{var_names}{Vector of variable names corresponding to `error_list`.}

\item{test_type}{Type of test, as a character}
}
\value{
Tidy `data.frame` summarizing the error output
}
\description{
Convert an error list into a tidy `data.frame`
}
\author{
Alexey Shiklomanov
}
