% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metutils.R
\name{sw2par}
\alias{sw2par}
\title{SW to PAR}
\usage{
sw2par(sw)
}
\arguments{
\item{sw}{shortwave radiation (W/m2 == J/m2/s)}
}
\value{
PAR W/m2
}
\description{
Solar Radiation to PPFD
}
\details{
Here the input is the total solar radiation
so to obtain in the PAR spectrum need to multiply by 0.486 From Campbell and Norman p151
This is based on the approximation that PAR is 0.45-0.50 of the total radiation
}
\author{
David LeBauer
}

