% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metutils.R
\name{sw2ppfd}
\alias{sw2ppfd}
\title{SW to PPFD}
\usage{
sw2ppfd(sw)
}
\arguments{
\item{SW}{CF surface_downwelling_shortwave_flux_in_air W/m2}
}
\value{
PPFD umol /m2 / s
}
\description{
CF Shortwave to PPFD
}
\details{
Cambell and Norman 1998 p 151, ch 10
}
\author{
David LeBauer
}

