% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metutils.R
\name{AirDens}
\alias{AirDens}
\title{Air Density}
\usage{
AirDens(pres, T, rv)
}
\arguments{
\item{pres}{air pressure (pascals)}

\item{T}{air temperature (Kelvin)}

\item{rv}{humidity}
}
\description{
estimate air density from pressure, temperature, and humidity
}
\author{
Mike Dietze
}

