% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cfmet.downscale.daily}
\alias{cfmet.downscale.daily}
\title{weachDT}
\usage{
cfmet.downscale.daily(dailymet, lat, output.dt = 1)
}
\arguments{
\item{X}{data table with climate variables}

\item{lat}{latitude (for calculating solar radiation)}

\item{output.dt}{output timestep}
}
\value{
weather file for input to BioGro and related crop growth functions
}
\description{
Simple, Fast Daily to Hourly Climate Downscaling
}
\details{
Based on weach family of functions but 5x faster than weachNEW,
and requiring metric units (temperature in celsius, windspeed in kph,
precip in mm, relative humidity as fraction)
}
\author{
David LeBauer
}

