% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.downscaling.R
\name{cfmet.downscale.daily}
\alias{cfmet.downscale.daily}
\title{daily to subdaily downscaling}
\usage{
cfmet.downscale.daily(dailymet, output.dt = 1, lat)
}
\arguments{
\item{dailymet}{data table with climate variables}

\item{output.dt}{output timestep}

\item{lat}{latitude (for calculating solar radiation)}
}
\value{
weather file for input to BioGro and related crop / ecosystem models
}
\description{
Simple, Fast Daily to Hourly Climate Downscaling
}
\details{
Based on weach family of functions but 5x faster than weachNEW,
and requiring metric units (temperature in celsius, windspeed in kph,
precip in mm, relative humidity as fraction). 
Derived from the weachDT function in the BioCro package.
}
\author{
David LeBauer
}

