% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cfmet.downscale.daily}
\alias{cfmet.downscale.daily}
\title{daily to subdaily downscaling}
\usage{
cfmet.downscale.daily(dailymet, output.dt = 1, lat = lat)
}
\arguments{
\item{dailymet}{data table with climate variables}

\item{lat}{latitude (for calculating solar radiation)}

\item{output.dt}{output timestep}
}
\value{
weather file for input to BioGro and related crop / ecosystem models
}
\description{
Simple, Fast Daily to Hourly Climate Downscaling
}
\details{
Based on weach family of functions but 5x faster than weachNEW,
and requiring metric units (temperature in celsius, windspeed in kph,
precip in mm, relative humidity as fraction).
Derived from the weachDT function in the BioCro package.
}
\author{
David LeBauer
}

