% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cfmet.downscale.time}
\alias{cfmet.downscale.time}
\title{Downscale CF met data}
\usage{
cfmet.downscale.time(cfmet, output.dt = 1, ...)
}
\arguments{
\item{cfmet}{data.table with CF variables generated by \code{\link{load.cfmet}}}

\item{output.dt}{time step for output}
}
\value{
downscaled result
}
\description{
Temporal downscaling of daily or subdaily met data
}
\author{
David LeBauer
}

