% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/temporal.downscaling.R
\name{cfmet.downscale.time}
\alias{cfmet.downscale.time}
\title{Downscale CF met data}
\usage{
cfmet.downscale.time(cfmet, output.dt = 1, lat = lat, ...)
}
\arguments{
\item{cfmet}{data.table with CF variables generated by \code{\link{load.cfmet}}}

\item{output.dt}{time step (hours) for output}
}
\value{
downscaled result
}
\description{
Temporal downscaling of daily or subdaily met data
}
\author{
David LeBauer
}

