% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_met_input_file}
\alias{check_met_input_file}
\title{Check a meteorology data file for compliance with the PEcAn standard}
\usage{
check_met_input_file(metfile, variable_table = pecan_standard_met_table)
}
\arguments{
\item{metfile}{Path of met file to check, as a scalar character.}

\item{variable_table}{`data.frame` linking standard names to their
units. Defaults to [pecan_standard_met_table].}
}
\value{
`TRUE` (invisibly) if successful, otherwise throw an informative error.
}
\description{
Check a meteorology data file for compliance with the PEcAn standard
}
\author{
Alexey Shiklomanov
}
