% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_met_input.R
\name{check_met_input_file}
\alias{check_met_input_file}
\title{Check a meteorology data file for compliance with the PEcAn standard}
\usage{
check_met_input_file(metfile, variable_table = pecan_standard_met_table,
  required_vars = variable_table \%>\% dplyr::filter(is_required) \%>\%
  dplyr::pull(cf_standard_name), warn_unknown = TRUE)
}
\arguments{
\item{metfile}{Path of met file to check, as a scalar character.}

\item{variable_table}{`data.frame` linking standard names to their
units. Must contain columns "cf_standard_name" and "units".
Default is [pecan_standard_met_table].}

\item{required_vars}{Character vector of required variables.
Defaults to variables marked as required in `variable_table`.}

\item{warn_unknown}{Logical. If `TRUE` (default), throw a warning
for variables not in `variable_table`. Otherwise, ignore unknown
variables.}
}
\value{
`data.frame` summarizing the results of the tests.
}
\description{
Check a meteorology data file for compliance with the PEcAn standard
}
\author{
Alexey Shiklomanov
}
