% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/closest_xy.R
\name{closest_xy}
\alias{closest_xy}
\title{closest_xy}
\usage{
closest_xy(slat, slon, infolder, infile)
}
\description{
Given latitude and longitude coordinates, find NARR x and y indices
}
\author{
Betsy Cowdery
}

