% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_angle.R
\name{cos_solar_zenith_angle}
\alias{cos_solar_zenith_angle}
\title{Cosine of solar zenith angle}
\usage{
cos_solar_zenith_angle(doy, lat, lon, dt, hr)
}
\arguments{
\item{doy}{Day of year}

\item{lat}{Latitude}

\item{lon}{Longitude}

\item{dt}{Timestep}

\item{hr}{Hours timestep}
}
\value{
`numeric(1)` of cosine of solar zenith angle
}
\description{
For explanations of formulae, see https://web.archive.org/web/20180307133425/http://www.itacanet.org/the-sun-as-a-source-of-energy/part-3-calculating-solar-angles/
}
\author{
Alexey Shiklomanov
}
