% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/met.process.R
\name{db.site.lat.lon}
\alias{db.site.lat.lon}
\title{db.site.lat.lon}
\usage{
db.site.lat.lon(site.id, con)
}
\arguments{
\item{site.id}{}

\item{con}{}
}
\description{
db.site.lat.lon
}
\author{
Betsy Cowdery
}

