% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download.Ameriflux.R
\name{download.Ameriflux}
\alias{download.Ameriflux}
\title{download.Ameriflux}
\usage{
download.Ameriflux(sitename, outfolder, start_date, end_date,
  overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}

\item{site}{the FLUXNET ID of the site to be downloaded, used as file name prefix.
The "SITE_ID" field in \href{http://ameriflux.lbl.gov/sites/site-list-and-pages/}{list of Ameriflux sites}}
}
\description{
Download Ameriflux L2 netCDF files
}
\author{
Josh Mantooth, Rob Kooper
}

