% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.Ameriflux.R
\name{download.Ameriflux}
\alias{download.Ameriflux}
\title{download.Ameriflux}
\usage{
download.Ameriflux(site, outfolder, start_date, end_date, overwrite = FALSE,
  verbose = FALSE)
}
\arguments{
\item{site}{the site to be downloaded, will be used as prefix as well}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}
}
\description{
Download Ameriflux L2 netCDF files
}
\author{
Josh Mantooth, Rob Kooper
}

