% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.CRUNCEP_Global.R
\name{download.CRUNCEP}
\alias{download.CRUNCEP}
\title{Download CRUNCEP data}
\usage{
download.CRUNCEP(outfolder, start_date, end_date, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, maxErrors = 10, sleep = 2,
  method = "ncss", ...)
}
\arguments{
\item{outfolder}{Directory where results should be written}

\item{start_date, end_date}{Range of years to retrieve. Format is YYYY-MM-DD,
but only the year portion is used and the resulting files always contain a full year of data.}

\item{lat.in}{site latitude in decimal degrees}

\item{lon.in}{site longitude in decimal degrees}

\item{overwrite}{logical. Download a fresh version even if a local file with the same name already exists?}

\item{verbose}{logical. Passed on to \code{\link[ncdf4]{ncvar_def}} and \code{\link[ncdf4]{nc_create}}
to control printing of debug info}

\item{maxErrors}{Maximum times to re-try folloing an error accessing netCDF data through THREDDS}

\item{sleep}{Wait time between attempts following a THREDDS or other error}

\item{method}{(string) Data access method. `opendap` (default)
attempts to directly access files via OpenDAP. `ncss` (NetCDF
subset) subsets the file on the server, downloads the subsetted
file to `tempfile` and then reads it locally. `opendap` is
faster when it works, but often fails because of server issues.
`ncss` can be much slower, but is more reliable.}

\item{...}{Other arguments, currently ignored}
}
\description{
Download and convert to CF CRUNCEP single grid point from MSTIMIP server using OPENDAP interface
}
\author{
James Simkins, Mike Dietze, Alexey Shiklomanov
}
