% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.CRUNCEP_Global.R
\name{download.CRUNCEP}
\alias{download.CRUNCEP}
\title{download.CRUNCEP}
\usage{
download.CRUNCEP(outfolder, start_date, end_date, lat, lon, overwrite = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{lon}{}
}
\description{
Download and conver to CF CRUNCEP single grid point from MSTIMIP server using OPENDAP interface
}
\author{
James Simkins
}

