% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ERA5.R
\name{download.ERA5.old}
\alias{download.ERA5.old}
\title{Download ERA 5 data}
\usage{
download.ERA5.old(
  outfolder,
  start_date,
  end_date,
  lat.in,
  lon.in,
  product_types = "all",
  overwrite = FALSE,
  reticulate_python = NULL,
  ...
)
}
\arguments{
\item{outfolder}{Directory where results should be written}

\item{start_date, end_date}{Range of years to retrieve. Format is
\code{YYYY-MM-DD}.}

\item{lat.in, lon.in}{Site coordinates, decimal degrees (numeric)}

\item{product_types}{Character vector of product types, or \code{"all"}.
Must be one or more of: \code{"reanalysis"}, \code{"ensemble members"},
\code{"ensemble mean"}, \code{"ensemble spread"}}

\item{overwrite}{Logical. If \code{FALSE} (default), skip any files with
the same target name (i.e. same variable) that already exist in
\code{outfolder}. If \code{TRUE}, silently overwrite existing files.}

\item{reticulate_python}{Path to Python binary for \code{reticulate}
(passed to \code{\link[reticulate:use_python]{reticulate::use_python()}}). If \code{NULL} (default), use
the system default.}

\item{...}{Currently unused. Allows soaking up additional arguments
to other methods.}
}
\value{
Character vector of file names containing raw, downloaded
data (invisibly)
}
\description{
Link to \href{https://confluence.ecmwf.int/display/CKB/ERA5+data+documentation}{full data documentation}.
}
\details{
Under the hood, this function uses the Python \code{cdsapi} module,
which can be installed via \code{pip} (\verb{pip install --user cdsapi}). The
module is accessed via the \code{reticulate} package.

Using the CDS API requires you to create a free account at
https://cds.climate.copernicus.eu. Once you have done that, you
will need to configure the CDS API on your local machine by
creating a \verb{$\{HOME\}/.cdsapi} file, as described
\href{https://cds.climate.copernicus.eu/api-how-to#install-the-cds-api-key}{here}.
}
\examples{
\dontrun{
files <- download.ERA5(
  "ERA5_output",
  start_date = "2010-01-01",
  end_date = "2010-02-01",
  lat.in = 45.5594,
  lon.in = -84.6738,
  product_types = "all"
)
}
}
\author{
Alexey Shiklomanov
}
