% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.ERA5.R
\name{download.ERA5}
\alias{download.ERA5}
\title{Download ERA 5 data}
\usage{
download.ERA5(outfolder, start_date, end_date, lat.in, lon.in,
  product_types = "all", reticulate_python = NULL, ...)
}
\arguments{
\item{outfolder}{Directory where results should be written}

\item{start_date, end_date}{Range of years to retrieve. Format is
`YYYY-MM-DD`.}

\item{lat.in, lon.in}{Site coordinates, decimal degrees (numeric)}

\item{product_types}{Character vector of product types, or `"all"`.
Must be one or more of: `"reanalysis"`, `"ensemble members"`,
`"ensemble mean"`, `"ensemble spread"`}

\item{reticulate_python}{Path to Python binary for `reticulate`
(passed to [reticulate::use_python()]). If `NULL` (default), use
the system default.}

\item{...}{Currently unused. Allows soaking up additional arguments
to other methods.}
}
\value{
Character vector of file names containing raw, downloaded
  data (invisibly)
}
\description{
Link to [full data documentation](https://confluence.ecmwf.int/display/CKB/ERA5+data+documentation).
}
\author{
Alexey Shiklomanov
}
