% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_download.R
\name{download.ERA5_cds}
\alias{download.ERA5_cds}
\title{ERA5_cds_annual_download}
\usage{
download.ERA5_cds(
  outfolder,
  start_date,
  end_date,
  extent,
  variables,
  auto.create.key = T,
  timeout = 36000
)
}
\arguments{
\item{outfolder}{Character: physical path where the ERA5 data are stored.}

\item{start_date}{character: the start date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)}

\item{end_date}{character: the end date of the data to be downloaded. Format is YYYY-MM-DD (will only use the year part of the date)}

\item{extent}{numeric: a vector of numbers contains the bounding box (formatted as xmin, xmax, ymin, ymax) to be downloaded.}

\item{variables}{character: a vector contains variables to be downloaded (e.g., c("2m_temperature","surface_pressure")).}

\item{auto.create.key}{Boolean: decide if we want to generate the CDS RC file if it doesn't exist, the default is TRUE.}

\item{timeout}{numeric: the maximum time (in seconds) allowed to download the data. The default is 36000 seconds.}
}
\value{
A vector containing file paths to the downloaded files.
}
\description{
This function helps to download the yearly ERA5 data based on the prescribed features using the CDS API.
}
\author{
Dongchen Zhang
}
