% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download.FACE}
\alias{download.FACE}
\title{download.FACE}
\usage{
download.FACE(sitename, outfolder, start_date, end_date, overwrite = FALSE)
}
\arguments{
\item{sitename}{}

\item{outfolder}{}

\item{start_year}{}

\item{end_year}{}
}
\description{
Download Raw FACE data from the internet
}
\author{
Betsy Cowdery
}

