% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.FACE.R
\name{download.FACE}
\alias{download.FACE}
\title{download.FACE}
\usage{
download.FACE(sitename, outfolder, start_date, end_date,
  overwrite = FALSE, method, ...)
}
\arguments{
\item{method}{Optional. Passed to download.file() function.  Use this to set custom programs such as ncftp to use when
downloading files from FTP sites}
}
\description{
Download Raw FACE data from the internet
}
\author{
Betsy Cowdery
}
