\name{download.GLDAS}
\alias{download.GLDAS}
\title{download.GLDAS}
\usage{
download.GLDAS(outfolder, start_date, end_date, site_id, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
  \item{outfolder}{}

  \item{start_date}{}

  \item{end_date}{}

  \item{site_id}{}

  \item{lat}{}

  \item{lon}{}
}
\description{
Download and convert single grid point GLDAS to CF single
grid point from hydro1.sci.gsfc.nasa.gov using OPENDAP
interface
}
\author{
Christy Rollinson
}

