% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.GLDAS.R
\name{download.GLDAS}
\alias{download.GLDAS}
\title{Download GLDAS data}
\usage{
download.GLDAS(outfolder, start_date, end_date, site_id, lat.in, lon.in,
  overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{lon.in}{}
}
\description{
Download and convert single grid point GLDAS to CF single grid point from hydro1.sci.gsfc.nasa.gov using OPENDAP interface
}
\author{
Christy Rollinson
}
