% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.MsTMIP_NARR.R
\name{download.MsTMIP_NARR}
\alias{download.MsTMIP_NARR}
\title{download.MsTMIP_NARR}
\usage{
download.MsTMIP_NARR(
  outfolder,
  start_date,
  end_date,
  site_id,
  lat.in,
  lon.in,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{outfolder}{location where output is stored}

\item{start_date}{YYYY-MM-DD}

\item{end_date}{YYYY-MM-DD}

\item{site_id}{BETY site id}

\item{lat.in}{latitude of site}

\item{lon.in}{longitude of site}

\item{overwrite}{overwrite existing files? Default is FALSE}

\item{verbose}{Default is FALSE, used in ncdf4::ncvar_def}

\item{...}{Other inputs}
}
\description{
Download and conver to CF NARR single grid point from MSTIMIP server using OPENDAP interface
}
\author{
James Simkins
}
