% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.MsTMIP_NARR.R
\name{download.MsTMIP_NARR}
\alias{download.MsTMIP_NARR}
\title{download.MsTMIP_NARR}
\usage{
download.MsTMIP_NARR(outfolder, start_date, end_date, site_id, lat.in,
  lon.in, overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{start_date}{YYYY-MM-DD}

\item{end_date}{YYYY-MM-DD}

\item{lat}{decimal degrees [-90, 90]}

\item{lon}{decimal degrees [-180, 180]}
}
\description{
Download and conver to CF NARR single grid point from MSTIMIP server using OPENDAP interface
}
\author{
James Simkins
}
