\name{download.NARR}
\alias{download.NARR}
\title{download.NARR}
\usage{
download.NARR(outfolder, start_date, end_date, overwrite = FALSE,
  verbose = FALSE, ...)
}
\arguments{
  \item{outfolder}{}

  \item{start_year}{}

  \item{end_year}{}
}
\description{
Download NARR files
}
\author{
Betsy Cowdery
}

