% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR.R
\name{download.NARR}
\alias{download.NARR}
\title{Download NARR files}
\usage{
download.NARR(outfolder, start_date, end_date, overwrite = FALSE,
  verbose = FALSE, ...)
}
\description{
Download NARR files
}
\author{
Betsy Cowdery
}
