% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download.NARR}
\alias{download.NARR}
\title{download.NARR}
\usage{
download.NARR(site.id, outfolder, start_date, end_date, overwrite = FALSE,
  verbose = FALSE, pkg, raw.host, dbparms, con)
}
\arguments{
\item{outfolder}{}

\item{start_year}{}

\item{end_year}{}
}
\description{
Download NARR files
}
\author{
??
}

