% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.NARR.R
\name{download.NARR}
\alias{download.NARR}
\title{download.NARR}
\usage{
download.NARR(outfolder, start_date, end_date, overwrite = FALSE)
}
\arguments{
\item{outfolder}{}

\item{start_year}{}

\item{end_year}{}
}
\description{
Download NARR files
}
\author{
Betsy Cowdery
}

