% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{download.NARR_site}
\alias{download.NARR_site}
\title{Download NARR time series for a single site}
\usage{
download.NARR_site(
  outfolder,
  start_date,
  end_date,
  lat.in,
  lon.in,
  overwrite = FALSE,
  verbose = FALSE,
  progress = TRUE,
  parallel = TRUE,
  ncores = if (parallel) parallel::detectCores() else NULL,
  ...
)
}
\arguments{
\item{outfolder}{Target directory for storing output}

\item{start_date}{Start date for met data}

\item{end_date}{End date for met data}

\item{lat.in}{Site latitude coordinate}

\item{lon.in}{Site longitude coordinate}

\item{overwrite}{Overwrite existing files?  Default=FALSE}

\item{verbose}{Turn on verbose output? Default=FALSE}

\item{progress}{Whether or not to show a progress bar.
Requires the `progress` package to be installed.}

\item{parallel}{Download in parallel? Default = TRUE}

\item{ncores}{Number of cores for parallel download. Default is
`parallel::detectCores()`}

\item{...}{further arguments, currently ignored}
}
\description{
Download NARR time series for a single site
}
\examples{

\dontrun{
download.NARR_site(tempdir(), "2001-01-01", "2001-01-12", 43.372, -89.907)
}


}
\author{
Alexey Shiklomanov
}
