% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NLDAS.R
\name{download.NLDAS}
\alias{download.NLDAS}
\title{Download NLDAS met data}
\usage{
download.NLDAS(
  outfolder,
  start_date,
  end_date,
  site_id,
  lat.in,
  lon.in,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{outfolder}{location of output}

\item{start_date}{desired start date YYYY-MM-DD}

\item{end_date}{desired end date YYYY-MM-DD}

\item{site_id}{site id (BETY)}

\item{lat.in}{latitude of site}

\item{lon.in}{longitude of site}

\item{overwrite}{overwrite existing files? Default is FALSE}

\item{verbose}{Turn on verbose output? Default=FALSE}

\item{...}{Other inputs}
}
\description{
Download and convert single grid point NLDAS to CF single grid point from hydro1.sci.gsfc.nasa.gov using OPENDAP interface
}
\author{
Christy Rollinson (with help from Ankur Desai)
}
