% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NOAA_GEFS.R
\name{download.NOAA_GEFS}
\alias{download.NOAA_GEFS}
\title{Script to launch NOAA download and temporal downscaling}
\usage{
download.NOAA_GEFS(
  site_id,
  sitename = NULL,
  username = "pecan",
  lat.in,
  lon.in,
  outfolder,
  start_date = Sys.Date(),
  end_date = start_date + lubridate::days(16),
  downscale = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{site_id, }{site identification number}

\item{downscale, }{logical specifying whether to downscale from 6-hr to 1-hr}

\item{overwrite, }{logical stating to overwrite any existing output_file}

\item{lat_list, }{vector of latitudes that correspond to site codes}

\item{lon_list, }{vector of longitudes that correspond to site codes}

\item{output_directory, }{directory where the model output will be save}
}
\value{
None
}
\description{
Script to launch NOAA download and temporal downscaling
}
\author{
Quinn Thomas ; modified by Katie Zarada
}
