% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NOAA_GEFS.R
\name{download.NOAA_GEFS}
\alias{download.NOAA_GEFS}
\title{Download NOAA GEFS Weather Data}
\usage{
download.NOAA_GEFS(outfolder, lat.in, lon.in, sitename,
  start_date = Sys.time(), end_date = (as.POSIXct(start_date, tz = "UTC") +
  lubridate::days(16)), overwrite = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{outfolder}{Directory where results should be written}

\item{sitename}{The unique ID given to each site. This is used as part of the file name.}

\item{start_date, }{end_date Range of dates/times to be downloaded (default assumed time of day is 0:00, midnight)}

\item{overwrite}{logical. Download a fresh version even if a local file with the same name already exists?}

\item{verbose}{logical.  Print additional debug information.  Passed on to functions in the netcdf4 package to provide debugging info.}

\item{...}{Other arguments, currently ignored}

\item{lat}{site latitude in decimal degrees}

\item{lon}{site longitude in decimal degrees}
}
\value{
A list of data frames is returned containing information about the data file that can be used to locate it later.  Each
data frame contains information about one file.

Download and convert to CF NOAA weather data
}
\examples{
\dontrun{
 download.NOAA_GEFS(outfolder="~/Working/results", lat.in= 45.805925, lon.in = -90.07961, sitename="US-WCr")
}

}
\references{
https://www.ncdc.noaa.gov/crn/measurements.html
Note that the temperature is measured in degrees C, but is converted at the station and downlaoded
in Kelvin.
}
\author{
Luke Dramko
}
