% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.PalEON.R
\name{download.PalEON}
\alias{download.PalEON}
\title{download.PalEON}
\usage{
download.PalEON(
  sitename,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{sitename}{sitename}

\item{outfolder}{desired output location}

\item{start_date}{desired start date YYYY-MM-DD}

\item{end_date}{desired end date YYYY-MM-DD}

\item{overwrite}{overwrite existing files? Default is FALSE}

\item{...}{Other inputs}
}
\description{
Download PalEON files
}
\author{
Betsy Cowdery
}
