\name{download.PalEON}
\alias{download.PalEON}
\title{download.PalEON}
\usage{
download.PalEON(sitename, outfolder, start_date, end_date, overwrite = FALSE,
  ...)
}
\arguments{
  \item{outfolder}{}

  \item{start_year}{}

  \item{end_year}{}
}
\description{
Download PalEON files
}
\author{
Betsy Cowdery
}

