% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaling_helper_functions.R
\name{downscale_repeat_6hr_to_hrly}
\alias{downscale_repeat_6hr_to_hrly}
\title{Downscale repeat to hourly}
\usage{
downscale_repeat_6hr_to_hrly(df, varName, hr = 1)
}
\arguments{
\item{df, }{dataframe of data to be downscaled (Longwave)}

\item{varName, }{variable names to be downscaled}

\item{hr, }{hour to downscale to- default is 1}
}
\value{
A dataframe of downscaled data
}
\description{
Downscale repeat to hourly
}
\author{
Laura Puckett
}
