% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_ShortWave_to_hrly.R
\name{downscale_ShortWave_to_hrly}
\alias{downscale_ShortWave_to_hrly}
\title{Downscale shortwave to hourly}
\usage{
downscale_ShortWave_to_hrly(df, lat, lon, hr = 1)
}
\arguments{
\item{lat, }{lat of site}

\item{lon, }{long of site}

\item{debiased, }{data frame of variables}

\item{time0, }{first timestep}

\item{time_end, }{last time step}

\item{output_tz, }{output timezone}
}
\value{
A dataframe of downscaled state variables
}
\description{
Downscale shortwave to hourly
}
\author{
Laura Puckett
}
