% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_met_process.R
\name{ERA5_met_process}
\alias{ERA5_met_process}
\title{Met Processes for ERA5 data}
\usage{
ERA5_met_process(settings, in.path, out.path, Write = FALSE)
}
\arguments{
\item{settings}{a multi-settings object}

\item{in.path}{met input path}

\item{out.path}{output path}

\item{Write}{if write into Bety database}
}
\value{
if Write is True then return input IDs with physical paths; if Write is False then return just physical paths of extracted ERA5 clim files.
}
\description{
Met Processes for ERA5 data
}
