% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERA5_met_process.R
\name{ERA5_met_process}
\alias{ERA5_met_process}
\title{Met Processes for ERA5 data}
\usage{
ERA5_met_process(settings, in.path, out.path, write.db = FALSE, write = TRUE)
}
\arguments{
\item{settings}{a multi-settings object}

\item{in.path}{met input path}

\item{out.path}{output path}

\item{write.db}{if write into Bety database}

\item{write}{if write the settings into pecan.xml file in the outdir of settings.}
}
\value{
if write.db is True then return input IDs with physical paths; if write.db is False then return just physical paths of extracted ERA5 clim files.
}
\description{
Met Processes for ERA5 data
}
\author{
Dongchen Zhang
}
