% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ERA5.R
\name{extract.nc.ERA5}
\alias{extract.nc.ERA5}
\title{ERA5_extract}
\usage{
extract.nc.ERA5(
  slat,
  slon,
  in.path,
  start_date,
  end_date,
  outfolder,
  in.prefix,
  newsite,
  ncores = 1,
  vars = NULL,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{slat}{numeric: vector of latitudes.}

\item{slon}{numeric: vector of longitudes.}

\item{in.path}{character: path to the directory containing the file to be inserted}

\item{start_date}{character: start date (in YYYY-MM-DD format).}

\item{end_date}{character: end date (in YYYY-MM-DD format).}

\item{outfolder}{character: Path to directory where nc files need to be saved.}

\item{in.prefix}{character: initial portion of the filename that does not vary by date.
Does not include directory; specify that as part of in.path.}

\item{newsite}{character: vector of site names. 
The length should match with that of slat and slon.}

\item{ncores}{numeric: the number of CPUs for the parallel compute. Default is 1.}

\item{vars}{character: names of variables to be extracted. If NULL all the variables will be
returned. Default is NULL.}

\item{overwrite}{Logical if files needs to be overwritten.}

\item{verbose}{Decide if we want to stop printing info.}

\item{...}{other inputs.}
}
\value{
a list of xts objects with all the variables for the requested years
}
\description{
ERA5_extract
}
\details{
For the list of variables check out the documentation at \url{
 https://confluence.ecmwf.int/display/CKB/ERA5+data+documentation#ERA5datadocumentation-Spatialgrid}
}
\examples{
\dontrun{
point.data <- ERA5_extract(sslat=40, slon=-120, years=c(1990:1995), vars=NULL)

 purrr::map(~xts::apply.daily(.x, mean))

}
}
