% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ERA5.R
\name{extract.nc.ERA5}
\alias{extract.nc.ERA5}
\title{ERA5_extract}
\usage{
extract.nc.ERA5(slat, slon, in.path, start_date, end_date, outfolder,
  in.prefix, newsite, vars = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{slat}{slatitude}

\item{slon}{slonitude}

\item{vars}{variables to be extarcted. If NULL all the variables will be returned.}

\item{years}{years to be extracted}
}
\value{
a list of xts objects with all the variables for the requested years
}
\description{
ERA5_extract
}
\details{
For the list of variables check out the documentation at \link{https://confluence.ecmwf.int/display/CKB/ERA5+data+documentation#ERA5datadocumentation-Spatialgrid}
}
\examples{
\dontrun{
point.data <- ERA5_extract(slat=40, slon=-120, years=c(1990:1995), vars=NULL)

 purrr::map(~xts::apply.daily(.x, mean))

}
}
