% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cf_variables_table.R
\name{get_cf_variables_table}
\alias{get_cf_variables_table}
\title{Retrieve the current CF variables table from \code{cfconventions.org}
and convert it into a \code{data.frame}}
\usage{
get_cf_variables_table(cf_url = build_cf_variables_table_url(57))
}
\arguments{
\item{cf_url}{URL of CF variables table XML. See also \link{build_cf_variables_table_url}.}
}
\value{
CF variables table, as a \code{tibble}
}
\description{
Retrieve the current CF variables table from \code{cfconventions.org}
and convert it into a \code{data.frame}
}
\author{
Alexey Shiklomanov
}
