% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_site.R
\name{get_NARR_thredds}
\alias{get_NARR_thredds}
\title{Retrieve NARR data using thredds}
\usage{
get_NARR_thredds(start_date, end_date, lat.in, lon.in, progress = TRUE,
  drop_outside = TRUE, parallel = TRUE, ncores = 1)
}
\arguments{
\item{start_date}{Start date for meteorology}

\item{end_date}{End date for meteorology}

\item{lat.in}{Latitude coordinate}

\item{lon.in}{Longitude coordinate}

\item{progress}{Whether or not to show a progress bar (default = `TRUE`).
Requires the `progress` package to be installed.}

\item{drop_outside}{Whether or not to drop dates outside of `start_date` to 
`end_date` range (default = `TRUE`).}

\item{parallel}{Download in parallel? Default = TRUE}

\item{ncores}{Number of cores for parallel download. Default is 
`parallel::detectCores()`}
}
\value{
`tibble` containing time series of NARR data for the given site
}
\description{
Retrieve NARR data using thredds
}
\examples{

\dontrun{
dat <- get_NARR_thredds("2008-01-01", "2008-01-15", 43.3724, -89.9071)
}

}
\author{
Alexey Shiklomanov
}
