% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.process.R
\name{met.process}
\alias{met.process}
\title{met.process}
\usage{
met.process(
  site,
  input_met,
  start_date,
  end_date,
  model,
  host = "localhost",
  dbparms,
  dir,
  spin = NULL,
  overwrite = FALSE,
  pfr_sites = NULL
)
}
\arguments{
\item{site}{Site info from settings file}

\item{input_met}{Which data source to process.}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{model}{model_type name}

\item{host}{Host info from settings file}

\item{dbparms}{database settings from settings file}

\item{dir}{directory to write outputs to}

\item{spin}{spin-up settings passed to model-specific met2model. List containing nyear (number of years of spin-up), nsample (first n years to cycle), and resample (TRUE/FALSE)}

\item{overwrite}{Whether to force met.process to proceed.


       `overwrite` may be a list with individual components corresponding to
       `download`, `met2cf`, `standardize`, and `met2model`. If it is instead a simple boolean,
       the default behavior for `overwrite=FALSE` is to overwrite nothing, as you might expect.
       Note however that the default behavior for `overwrite=TRUE` is to overwrite everything
       *except* raw met downloads. I.e., it corresponds to:

       list(download = FALSE, met2cf = TRUE, standardize = TRUE,  met2model = TRUE)
 List of `url`, `username`, `password`}

\item{pfr_sites}{Optional permafrost and soil-property table for SIPNET.
Contains `index`, `is_permafrost`, and `sand_pct`. May contain multiple
sites; the current site's row is selected by matching `index` to
`site$id`.}
}
\description{
met.process
}
\author{
Elizabeth Cowdery, Michael Dietze, Ankur Desai, James Simkins, Ryan Kelly
}
