% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.process.stage.R
\name{met.process.stage}
\alias{met.process.stage}
\title{met.process.stage}
\usage{
met.process.stage(input.id, raw.id, con)
}
\arguments{
\item{input.id}{bety db for input format}

\item{raw.id}{format id for the raw met data}

\item{con}{database connection}
}
\description{
met.process.stage
}
\author{
Elizabeth Cowdery
}
