% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2CF.AmerifluxLBL.R
\name{met2CF.AmerifluxLBL}
\alias{met2CF.AmerifluxLBL}
\title{met2CF.AmerifluxLBL}
\usage{
met2CF.AmerifluxLBL(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  format,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{in.path}{location on disk where inputs are stored}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{format}{is data frame or list with elements as described below
The AmerifluxLBL format is Bety record 5000000002
which could be returned from PEcAn.DB::query.format.vars(format.id=5000000002)
format is output from db/R/query.format.vars, and should have:
  REQUIRED:
  format$lat = latitude of site
  format$lon = longitude of site
  format$header = number of lines of header
  format$vars is a data.frame with lists of information for each variable to read, at least airT is required
    format$vars$input_name = Name in CSV file
    format$vars$input_units = Units in CSV file
    format$vars$bety_name = Name in BETY
  OPTIONAL:
  format$na.strings = list of missing values to convert to NA, such as -9999
  format$skip = lines to skip excluding header
  format$vars$column_number = Column number in CSV file (optional, will use header name first)
Columns with NA for bety variable name are dropped. 
Units for datetime field are the lubridate function that will be used to parse the date (e.g. \code{ymd_hms} or \code{mdy_hm}).}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should ouput of function be extra verbose}

\item{...}{further arguments, currently ignored}
}
\description{
Get meteorology variables from Ameriflux LBL and convert to netCDF CF format
}
\author{
Ankur Desai
}
