\name{met2CF.csv}
\alias{met2CF.csv}
\title{met2CF.csv}
\arguments{
  \item{in.path}{}

  \item{in.prefix}{}

  \item{outfolder}{}

  \item{format}{data frame or list with elements orig,
  bety, units for the original variable name, bety variable
  name, and original units}
}
\description{
  met2CF.csv
}
\author{
  Mike Dietze
}

