% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2cf.ERA5.R
\name{met2CF.ERA5}
\alias{met2CF.ERA5}
\title{met2cf.ERA5}
\usage{
met2CF.ERA5(lat, long, start_date, end_date, sitename, outfolder, out.xts,
  overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{lat}{latitude}

\item{long}{longitude}

\item{start_date}{start date}

\item{end_date}{end date}

\item{sitename}{The name of the site used for making the identifier.}

\item{outfolder}{Path to directory where nc files need to be saved.}

\item{out.xts}{Output of the extract.nc.ERA5 function which is a list of time series of met variables for each ensemble member.}

\item{overwrite}{Logical if files needs to be overwritten.}

\item{verbose}{Logical flag defining if ouput of function be extra verbose.}
}
\value{
list of dataframes
}
\description{
met2cf.ERA5
}
