% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/met2CF.FACE.R
\name{met2CF.FACE}
\alias{met2CF.FACE}
\title{met2CF.FACE}
\usage{
met2CF.FACE(in.path, in.prefix, outfolder, start_date, end_date)
}
\arguments{
\item{in.path}{}

\item{in.prefix}{}

\item{outfolder}{}

\item{convert}{FACE files to CF files}
}
\description{
met2CF.FACE
}
\author{
Elizabeth Cowdery
}

