% Generated by roxygen2 (4.0.2): do not edit by hand
\name{met2CF.NARR}
\alias{met2CF.NARR}
\title{met2CF.NARR}
\usage{
met2CF.NARR(in.path, in.prefix, outfolder)
}
\arguments{
\item{in.path}{}

\item{in.prefix}{}

\item{outfolder}{}

\item{convert}{NARR files to CF files}
}
\description{
met2CF.NARR
}
\author{
Elizabeth Cowdery, Rob Kooper
}

