% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metgapfill.R
\name{metgapfill}
\alias{metgapfill}
\title{Take an Ameriflux NetCDF file
Fill missing met values using MDS approach using MPI-BGC REddyProc library
Currently
Future version: Choose which variables to gap fill
Future version will first downscale and fill with NARR, then REddyProc}
\usage{
metgapfill(in.path, in.prefix, outfolder, start_date, end_date, lst = 0,
  overwrite = FALSE, verbose = FALSE)
}
\arguments{
\item{in.path}{location on disk where inputs are stord}

\item{in.prefix}{prefix of input and output files}

\item{outfolder}{location on disk where outputs will be stored}

\item{start_date}{the start date of the data to be downloaded (will only use the year part of the date)}

\item{end_date}{the end date of the data to be downloaded (will only use the year part of the date)}

\item{lst}{is timezone offset from UTC, if timezone is available in time:units atribute in file, it will use that, default is to assume UTC}

\item{overwrite}{should existing files be overwritten}

\item{verbose}{should the function be very verbose}
}
\description{
Take an Ameriflux NetCDF file
Fill missing met values using MDS approach using MPI-BGC REddyProc library
Currently
Future version: Choose which variables to gap fill
Future version will first downscale and fill with NARR, then REddyProc
}
\author{
Ankur Desai
}

