% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_merge.R
\name{nc.merge}
\alias{nc.merge}
\title{nc.merge}
\usage{
nc.merge(
  outfolder,
  in.path,
  in.prefix,
  start_date,
  end_date,
  upscale = FALSE,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{outfolder}{- directory where output will be stored}

\item{in.path}{- path of coarse model (e.g. GCM output)}

\item{in.prefix}{- prefix of model string as character (e.g. IPSL.r1i1p1.rcp85)}

\item{start_date}{- yyyy-mm-dd}

\item{end_date}{- yyyy-mm-dd}

\item{upscale}{- Upscale can either be set for FALSE (leave alone) or to the temporal resolution you want to aggregate to}

\item{overwrite}{logical: replace output file if it already exists?}

\item{verbose}{logical: should \code{\link[ncdf4:ncdf4-package]{ncdf4}}
functions print debugging information as they run?}

\item{...}{further arguments, currently ignored}
}
\description{
This is the 1st function for the tdm (Temporally Downscale Meteorology) workflow. The nc2dat.train function
             parses multiple netCDF files into one central training data file called 'dat.train_file'. This netCDF
             file will be used to generate the subdaily models in the next step of the workflow, generate.subdaily.models(). 
             It is also called in tdm_predict_subdaily_met which is the final step of the tdm workflow.
}
\details{
nc.merge
Parses multiple netCDF files into one central document for temporal downscaling procedure
}
\seealso{
Other tdm - Temporally Downscale Meteorology: 
\code{\link{gen.subdaily.models}()},
\code{\link{lm_ensemble_sims}()},
\code{\link{model.train}()},
\code{\link{predict_subdaily_met}()},
\code{\link{save.betas}()},
\code{\link{save.model}()},
\code{\link{subdaily_pred}()},
\code{\link{temporal.downscale.functions}()}
}
\author{
James Simkins, Christy Rollinson
}
\concept{tdm - Temporally Downscale Meteorology}
