% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_predict_subdaily_met.R
\name{predict_subdaily_met}
\alias{predict_subdaily_met}
\title{predict_subdaily_met}
\usage{
predict_subdaily_met(outfolder, in.path, in.prefix, lm.models.base,
  dat.train_file, start_date, end_date, cores.max = 12, n.ens = 3,
  resids = FALSE, parallel = FALSE, n.cores = NULL, overwrite = FALSE,
  verbose = FALSE)
}
\arguments{
\item{outfolder}{- directory where output file will be stored}

\item{in.path}{- path to model dataset you wish to temporally downscale}

\item{in.prefix}{- prefix of model dataset, i.e. if file is GFDL.CM3.rcp45.r1i1p1.2006 the prefix is 'GFDL.CM3.rcp45.r1i1p1'}

\item{lm.models.base}{- path to linear regression model folder from 3_gen_subdaily}

\item{dat.train_file}{- location of train_data file}

\item{start_date}{- yyyy-mm-dd}

\item{end_date}{- yyyy-mm-dd}

\item{cores.max}{- 12}

\item{n.ens}{- integer selecting number of hourly ensemble members}

\item{resids}{- logical stating whether to pass on residual data or not}

\item{parallel}{- logical stating whether to run temporal_downscale_functions.R in parallel}

\item{n.cores}{- deals with parallelization}
}
\description{
This is the main function of the tdm family workflow. This function predicts subdaily meteorology
             from daily means using a linear regression modeling approach. It takes a dataset with 
             daily resolution and temporally downscales it to hourly resolution using the statistics
             generated by gen.subdaily.models(). It references the predict.subdaily.function
             located in lm_ensemble_sims() which uses a linear regression based approach to downscale.
             We generate multiple ensembles of possible hourly values dictated from the models and betas 
             generated in gen.subdaily.models. Each ensemble member is saved as a netCDF file
             in CF conventions and these files are ready to be used in the general PEcAn workflow.
}
\details{
Predict Subdaily Meteorology
Predict Subdaily Meteorology based off of statistics created in gen.subdaily.models()
}
\examples{
\dontrun{
library(PEcAn.data.atmosphere)
outfolder = '~/Downscaled_GCM'
in.path = '~/raw_GCM'
in.prefix = 'GFDL'
lm.models.base = 'sf_scratch/US-WCr'
dat.train_file = 'Training_data/US-WCr_dat.train.nc'
start_date = '2010-01-01'
end_date = '2014-12-31'
cores.max = 12
n.ens = 3}
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{lm_ensemble_sims}},
  \code{\link{model.train}}, \code{\link{nc.merge}},
  \code{\link{save.betas}}, \code{\link{save.model}},
  \code{\link{subdaily_pred}},
  \code{\link{temporal.downscale.functions}}
}
\author{
Christy Rollinson, James Simkins
}
