% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEFS_helper_functions.R
\name{process_gridded_noaa_download}
\alias{process_gridded_noaa_download}
\title{Extract and temporally downscale points from downloaded grid files}
\usage{
process_gridded_noaa_download(
  lat_list,
  lon_list,
  site_id,
  downscale,
  overwrite,
  forecast_date,
  forecast_time,
  model_name,
  model_name_ds,
  model_name_raw,
  output_directory
)
}
\arguments{
\item{output_directory}{}
}
\value{

}
\description{
Extract and temporally downscale points from downloaded grid files
}
\examples{

}
