% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.register.R
\name{read.register}
\alias{read.register}
\title{read.register}
\usage{
read.register(register.xml, con)
}
\arguments{
\item{register.xml}{path of xml file}

\item{con}{betydb connection}
}
\author{
Betsy Cowdery
}

