% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.NARR_thredds.R
\name{read_narr_var}
\alias{read_narr_var}
\title{Read a specific variable from a NARR NetCDF file}
\usage{
read_narr_var(nc, xy, variable, unit, flx, pb = NULL)
}
\arguments{
\item{nc}{`ncdf4` connection object}

\item{xy}{Vector length 2 containing NARR coordinates}

\item{variable}{NARR name of variable to retrieve}

\item{unit}{Output unit of variable to retrieve}

\item{flx}{(Logical) If `TRUE`, format for `flx` variables. Otherwise, 
format for `sfc` variables. See [narr_flx_vars].}

\item{pb}{Progress bar R6 object (default = `NULL`)}
}
\description{
Read a specific variable from a NARR NetCDF file
}
\author{
Alexey Shiklomanov
}
