% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metutils.R
\name{rh2qair}
\alias{rh2qair}
\alias{rh2rv}
\title{RH to SH}
\usage{
rh2qair(rh, T)
}
\arguments{
\item{rh}{relative humidity (proportion, not %)}

\item{T}{absolute temperature (Kelvin)}
}
\description{
converts relative humidity to specific humidity
}
\author{
Mike Dietze
}

