% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_save_model.R
\name{save.model}
\alias{save.model}
\title{save.model}
\usage{
save.model(model.out, model, outfile)
}
\arguments{
\item{model.out}{list linear regression model output}

\item{model}{name of the layer of model to save (e.g. 'model' or 'model.resid')}

\item{outfile}{location where output will be stored}
}
\description{
Function to save models as a .nc file. This is utilized in
             gen.subdaily.models() when linear regression models are created
}
\details{
TDM Save Models
Saves models that are created during gen.subdaily.models()
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{lm_ensemble_sims}},
  \code{\link{model.train}}, \code{\link{nc.merge}},
  \code{\link{predict_subdaily_met}},
  \code{\link{save.betas}}, \code{\link{subdaily_pred}},
  \code{\link{temporal.downscale.functions}}
}
\author{
Christy Rollinson, James Simkins
}
\concept{tdm - Temporally Downscale Meteorology}
