% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/met.process.R
\name{site_from_tag}
\alias{site_from_tag}
\title{site_from_tag}
\usage{
site_from_tag(sitename, tag)
}
\arguments{
\item{sitename}{}

\item{tag}{}
}
\description{
site_from_tag
}
\author{
Betsy Cowdery

Function to find the site code for a specific tag
Example:
  sitename = "Rhinelander Aspen FACE Experiment (FACE-RHIN)"
  tag = "FACE"
  site_from_tag(sitename,tag) = "RHIN"
Requires that site names be set up specifically with (tag-sitecode) - this may change
}

