% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_temporal_downscale_functions.R
\name{temporal.downscale.functions}
\alias{temporal.downscale.functions}
\title{temporal.downscale.functions}
\usage{
temporal.downscale.functions()
}
\arguments{
\item{dat.train}{- training data generated by tdm_nc2dat.train.R}

\item{n.beta}{- number of betas to generate}

\item{path.out}{- path to where the training models & betas will be stored}

\item{resids}{- whether or not to propogate residuals, set to FALSE}

\item{parallel}{- whether or not to run in parallel. this is a feature
still being worked on, set to FALSE}

\item{n.cores}{- number of cores to use parallel processing on, set to NULL}

\item{day.window}{- number of days surrounding current day we want to pull
statistics from}

\item{seed}{- allows this to be reproducible}
}
\description{
This script contains the functions that do the heavy lifting in gen.subdaily.models
             and predict.subdaily.workflow. These functions actually generate the models and betas
             from the dat.train_file and save them in the output file.
}
\details{
Met variable functions that are called in gen_subdaily_models and predict_subdaily_met
}
\seealso{
Other tdm - Temporally Downscale Meteorology: \code{\link{gen.subdaily.models}},
  \code{\link{nc2dat.train}},
  \code{\link{predict.subdaily.function}},
  \code{\link{predict.subdaily.workflow}}
}
\author{
Christy Rollinson, James Simkins
}
