% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdm_temporal_downscale_functions.R
\name{temporal.downscale.functions}
\alias{temporal.downscale.functions}
\title{temporal_downscale_functions}
\usage{
temporal.downscale.functions(
  dat.train,
  n.beta,
  day.window,
  resids = FALSE,
  parallel = FALSE,
  n.cores = NULL,
  seed = format(Sys.time(), "\%m\%d"),
  outfolder,
  print.progress = FALSE,
  ...
)
}
\arguments{
\item{dat.train}{- training data generated by tdm_nc2dat.train.R}

\item{n.beta}{- number of betas to generate}

\item{day.window}{- number of days surrounding current day we want to pull
statistics from}

\item{resids}{- whether or not to propagate residuals, set to FALSE}

\item{parallel}{- whether or not to run in parallel. this is a feature
still being worked on, set to FALSE}

\item{n.cores}{- number of cores to use parallel processing on, set to NULL}

\item{seed}{- allows this to be reproducible}

\item{outfolder}{= where the output should be stored}

\item{print.progress}{- print progress of model generation?}

\item{...}{further arguments, currently ignored}
}
\description{
This function contains the functions that do the heavy lifting in gen.subdaily.models()
             and predict.subdaily.workflow(). Individual variable functions actually generate the models 
             and betas from the dat.train_file and save them in the output file. save.model() and 
             save.betas() are helper functions that save the linear regression model output to a 
             specific location. In the future, we should only save the data that we actually use from the 
             linear regression model because this is a large file. predict.met() is called from
             predict.subdaily.workflow() and references the linear regression model output to 
             predict the ensemble data.
}
\details{
Temporal Downscale Functions
 Met variable functions that are called in gen.subdaily.models and predict.subdaily.workflow
}
\seealso{
Other tdm - Temporally Downscale Meteorology: 
\code{\link{gen.subdaily.models}()},
\code{\link{lm_ensemble_sims}()},
\code{\link{model.train}()},
\code{\link{nc.merge}()},
\code{\link{predict_subdaily_met}()},
\code{\link{save.betas}()},
\code{\link{save.model}()},
\code{\link{subdaily_pred}()}
}
\author{
Christy Rollinson, James Simkins
}
\concept{tdm - Temporally Downscale Meteorology}
