% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upscale_met.R
\name{upscale_met}
\alias{upscale_met}
\title{upscale_met}
\usage{
upscale_met(outfolder, input_met, site.id, reso = 6, overwrite = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{input_met}{- met dataset}

\item{reso}{- numeric value in HOURS that is the desired resolution for the dataset to be upscaled to
# might need to change this for monthly/yearly option later}
}
\description{
upscale_met upscales the temporal resolution of a dataset
}
\author{
James Simkins
}

